refactor(desktop): split ChatPane.tsx into a focused module structure (#313)
* refactor(desktop): extract ChatPane types/constants/helpers (Phase 1/5)

Pure data + pure functions moved out of the 11.9k-line ChatPane.tsx
into focused sibling modules under ./ChatPane/. No runtime behavior
change; the orchestrator imports them back.

- types.ts        — all type/interface declarations
- constants.ts    — module-level constants, EMPTY_* sentinels,
                    storage keys, presets
- helpers.ts      — pure functions (slash/mention parsers, model
                    label formatters, time labels, attachment
                    type checks)


* refactor(desktop): extract ChatPane leaf components (Phase 2/5)

- skeletons.tsx
- AttachmentList.tsx (with AttachmentImageThumb)
- ImageAttachmentPreviewModal.tsx
- ArtifactBrowserModal.tsx (with OutputArtifactIcon)
- AssistantTurn/status.tsx (LiveStatusLine, Ellipsis, Cursor,
  TraceTimelineStepEntry, ExecutionTimelineThinkingEntry)

ArtifactBrowserModal export preserved on the public surface via
re-export from ChatPane.tsx.


* refactor(desktop): extract ChatPane mid-tier components (Phase 3/5)

- ChatHeader.tsx
- QueuedSessionInputRail.tsx
- AssistantTurn/{ActionsMenu,Outputs,MemoryProposals,TraceStepGroup}.tsx
- AssistantTurn/index.tsx — orchestrates the cluster


* refactor(desktop): extract Composer cluster (Phase 4/5)

- Composer/ModelCombobox.tsx
- Composer/ThinkingValueSelect.tsx
- Composer/index.tsx — Composer with its ~50-prop ComposerProps
  surface preserved


* refactor(desktop): finalize ChatPane split — orchestrator becomes index (Phase 5/5)

- UserTurn.tsx
- ConversationTurns.tsx (re-exported from index for external consumers)
- ChatPane.tsx → ChatPane/index.tsx; sibling imports normalized

External imports of `@/components/panes/ChatPane` resolve to
ChatPane/index.tsx unchanged. No runtime behavior delta.

Test file ChatPane.test.mjs source path updated to read from
ChatPane/index.tsx. Many assertions now fail because their
patterns live in sibling files; this is expected — the test scope
is unchanged from before this PR.


---------
