+++ desktop/src/components/panes/ChatPane/ChatHeader.tsx
@@ -1,0 +1,94 @@
+import { Clock3, Inbox, MessageCircle } from "lucide-react";
+import { Button } from "@/components/ui/button";
+import { AgentAvatar } from "@/components/ui/agent-avatar";
+import { StatusDot } from "@/components/ui/status-dot";
+
+interface ChatHeaderProps {
+  agentName: string;
+  workspace: WorkspaceRecordPayload | null;
+  subtitle?: string;
+  onReturnToMainSession?: () => void;
+  onOpenSessions?: () => void;
+  onOpenInbox?: () => void;
+  inboxUnreadCount: number;
+  onOpenAutomations?: () => void;
+}
+
+export function ChatHeader({
+  agentName,
+  workspace,
+  subtitle,
+  onReturnToMainSession,
+  onOpenSessions,
+  onOpenInbox,
+  inboxUnreadCount,
+  onOpenAutomations,
+}: ChatHeaderProps) {
+  const seed = workspace?.id ?? agentName ?? "default";
+
+  return (
+    <div className="flex items-center justify-between gap-3">
+      <div className="flex min-w-0 flex-1 items-center gap-2">
+        <AgentAvatar seed={seed} size="sm" />
+        <div className="flex min-w-0 flex-col">
+          <span className="truncate text-sm font-medium text-foreground">
+            {agentName}
+          </span>
+          {subtitle ? (
+            <span className="truncate text-[11px] leading-tight text-muted-foreground">
+              {subtitle}
+            </span>
+          ) : null}
+        </div>
+      </div>
+
+      <div className="flex shrink-0 items-center gap-0.5">
+        {onOpenSessions ? (
+          <Button
+            type="button"
+            variant="ghost"
+            size="icon-sm"
+            onClick={() => onOpenSessions()}
+            aria-label="Sessions"
+            className="text-muted-foreground hover:text-foreground"
+          >
+            <MessageCircle className="size-4" />
+          </Button>
+        ) : null}
+
+        {onOpenInbox ? (
+          <Button
+            type="button"
+            variant="ghost"
+            size="icon-sm"
+            onClick={() => onOpenInbox()}
+            aria-label="Inbox"
+            className="relative text-muted-foreground hover:text-foreground"
+          >
+            <Inbox className="size-4" />
+            {inboxUnreadCount > 0 ? (
+              <StatusDot
+                variant="destructive"
+                size="sm"
+                className="absolute right-1 top-1 border border-card"
+              />
+            ) : null}
+          </Button>
+        ) : null}
+
+        {onOpenAutomations ? (
+          <Button
+            type="button"
+            variant="ghost"
+            size="icon-sm"
+            onClick={() => onOpenAutomations()}
+            aria-label="Automations"
+            className="text-muted-foreground hover:text-foreground"
+          >
+            <Clock3 className="size-4" />
+          </Button>
+        ) : null}
+      </div>
+    </div>
+  );
+}
