+++ desktop/src/components/panes/ChatPane/index.tsx
@@ -113,319 +113,132 @@ import {
   type WorkspaceFileEntry,
 } from "@/lib/workspaceFiles";
 import { useWorkspaceSelection } from "@/lib/workspaceSelection";
-import * as modelCatalog from "../../../shared/model-catalog.js";
+import * as modelCatalog from "../../../../shared/model-catalog.js";
+import {
+  type ChatAttachment,
+  type ChatPaneVariant,
+  type ChatAssistantSegment,
+  type ChatMessage,
+  type QueuedSessionInputStatus,
+  type QueuedSessionInput,
+  type QueuedSessionInputPreviewDescriptor,
+  type ComposerInputRecallSnapshot,
+  type PendingOptimisticUserMessage,
+  type ChatSerializedQuotedSkillBlock,
+  type ChatTraceStepStatus,
+  type ChatTraceStep,
+  type ChatExecutionTimelineItem,
+  type PendingLocalAttachmentFile,
+  type PendingExplorerAttachmentFile,
+  type PendingAttachment,
+  type AttachmentListItem,
+  type ImageAttachmentPreviewState,
+  type ChatModelOption,
+  type ChatModelOptionGroup,
+  type ChatComposerSlashCommandOption,
+  type ChatComposerQuotedSkillItem,
+  type ChatComposerMentionItem,
+  type StreamTelemetryEntry,
+  type ArtifactBrowserFilter,
+} from "./types";
+import {
+  MAIN_SESSION_EVENT_BATCH_HEADER,
+  BACKGROUND_DELIVERY_RETRY_STATUS_MESSAGE,
+  EMPTY_ATTACHMENTS,
+  EMPTY_SEGMENTS,
+  EMPTY_EXECUTION_ITEMS,
+  EMPTY_OUTPUTS,
+  EMPTY_MEMORY_PROPOSALS,
+  STREAM_ATTACH_PENDING,
+  STREAM_TELEMETRY_LIMIT,
+  TOOL_TRACE_TERMINAL_PHASES,
+  CHAT_AUTO_SCROLL_THRESHOLD_PX,
+  CHAT_HISTORY_PAGE_SIZE,
+  CHAT_HISTORY_TOP_LOAD_THRESHOLD_PX,
+  COMPOSER_FOOTER_GAP_PX,
+  COMPOSER_FULL_MODEL_CONTROL_WIDTH_PX,
+  COMPOSER_FULL_THINKING_CONTROL_WIDTH_PX,
+  COMPOSER_FULL_PROVIDER_SETUP_WIDTH_PX,
+  COMPOSER_COMPACT_MODEL_CONTROL_MAX_WIDTH_PX,
+  COMPOSER_COMPACT_THINKING_CONTROL_MIN_WIDTH_PX,
+  COMPOSER_COMPACT_THINKING_CONTROL_MAX_WIDTH_PX,
+  CHAT_MODEL_STORAGE_KEY,
+  CHAT_THINKING_STORAGE_KEY,
+  CHAT_MODEL_USE_RUNTIME_DEFAULT,
+  CHAT_SERIALIZED_SKILL_COMMAND_PATTERN,
+  QUEUED_MESSAGES_PREVIEW_EVENT,
+  LEGACY_UNAVAILABLE_CHAT_MODELS,
+  DEPRECATED_CHAT_MODELS,
+  CHAT_MODEL_PRESETS,
+  RUNTIME_MODEL_CAPABILITY_ALIASES,
+  MENTION_TOKEN_PATTERN,
+} from "./constants";
+import {
+  initialBrowserState,
+  attachmentLooksLikeImage,
+  pendingAttachmentIsImage,
+  supportsImageInput,
+  imageInputUnsupportedMessage,
+  parseSerializedQuotedSkillPrompt,
+  appendComposerPrefillText,
+  buildComposerSlashCommandOptions,
+  findActiveSlashCommandRange,
+  removeSlashCommandText,
+  findActiveMentionRange,
+  replaceMentionText,
+  injectMentionLinks,
+  displayModelLabel,
+  compactComposerModelLabel,
+  chatMessageTimeLabel,
+  inputIdFromMessageId,
+  inputIdFromHistoryMessage,
+  historyMessagesInDisplayOrder,
+  turnInputIdsFromHistoryMessages,
+} from "./helpers";
+import {
+  HistoryRestoreSkeleton,
+  IntegrationErrorBanner,
+} from "./skeletons";
+import {
+  AttachmentList,
+  formatAttachmentSize,
+} from "./AttachmentList";
+import { ImageAttachmentPreviewModal } from "./ImageAttachmentPreviewModal";
+import {
+  ArtifactBrowserModal,
+  OutputArtifactIcon,
+  outputBrowserFilterForOutput,
+  outputChangeLabel,
+  outputKindLabel,
+  outputMetadataNumber,
+  outputMetadataString,
+  outputSecondaryLabel,
+  sortOutputs,
+  sortOutputsLatestFirst,
+  dedupeOutputsForDisplay,
+} from "./ArtifactBrowserModal";
+import { ChatHeader } from "./ChatHeader";
+import { QueuedSessionInputRail } from "./QueuedSessionInputRail";
+import { AssistantTurn } from "./AssistantTurn";
+import { Composer } from "./Composer";
+import { UserTurn } from "./UserTurn";
+import { ConversationTurns } from "./ConversationTurns";
 
-type ChatAttachment = SessionInputAttachmentPayload;
-type ChatPaneVariant = "default" | "onboarding";
-const MAIN_SESSION_EVENT_BATCH_HEADER =
-  "[Holaboss Main Session Event Batch v1]";
-const BACKGROUND_DELIVERY_RETRY_STATUS_MESSAGE =
-  "Background update delayed. Retrying automatically.";
-
-export type ChatAssistantSegment =
-  | {
-      kind: "execution";
-      items: ChatExecutionTimelineItem[];
-    }
-  | {
-      kind: "output";
-      text: string;
-      tone?: "default" | "error";
-    };
-
-const EMPTY_ATTACHMENTS: ChatAttachment[] = [];
-const EMPTY_SEGMENTS: ChatAssistantSegment[] = [];
-const EMPTY_EXECUTION_ITEMS: ChatExecutionTimelineItem[] = [];
-const EMPTY_OUTPUTS: WorkspaceOutputRecordPayload[] = [];
-const EMPTY_MEMORY_PROPOSALS: MemoryUpdateProposalRecordPayload[] = [];
-
-export interface ChatMessage {
-  id: string;
-  role: "user" | "assistant";
-  text: string;
-  tone?: "default" | "error";
-  createdAt?: string;
-  attachments?: ChatAttachment[];
-  segments?: ChatAssistantSegment[];
-  executionItems?: ChatExecutionTimelineItem[];
-  outputs?: WorkspaceOutputRecordPayload[];
-  memoryProposals?: MemoryUpdateProposalRecordPayload[];
-}
-
-type QueuedSessionInputStatus = "queued" | "sending";
-
-interface QueuedSessionInput {
-  inputId: string;
-  sessionId: string;
-  workspaceId: string;
-  text: string;
-  createdAt: string;
-  attachments: ChatAttachment[];
-  status: QueuedSessionInputStatus;
-}
-
-interface QueuedSessionInputPreviewDescriptor {
-  text: string;
-  createdAt?: string;
-  attachments?: ChatAttachment[];
-  status: QueuedSessionInputStatus;
-}
-
-interface ComposerInputRecallSnapshot {
-  workspaceId: string;
-  text: string;
-  at: number;
-}
-
-interface PendingOptimisticUserMessage {
-  localMessageId: string;
-  inputId?: string | null;
-  sessionId: string;
-  workspaceId: string;
-  message: ChatMessage;
-}
-
-declare global {
-  interface Window {
-    __holabossQueuedMessagesPreviewState?: QueuedSessionInputPreviewDescriptor[];
-    __holabossDevQueuedMessagesPreview?: {
-      single: (text?: string) => void;
-      multiple: () => void;
-      clear: () => void;
-      set: (
-        entries:
-          | string
-          | Array<string | Partial<QueuedSessionInputPreviewDescriptor>>,
-      ) => void;
-      get: () => QueuedSessionInputPreviewDescriptor[];
-    };
-  }
-}
-
-interface ChatSerializedQuotedSkillBlock {
-  skillIds: string[];
-  body: string;
-}
-
-function initialBrowserState(space: BrowserSpaceId): BrowserTabListPayload {
-  return {
-    space,
-    activeTabId: "",
-    tabs: [],
-    tabCounts: {
-      user: 0,
-      agent: 0,
-    },
-    sessionId: null,
-    lifecycleState: null,
-    controlMode: "none",
-    controlSessionId: null,
-  };
-}
-
-type ChatTraceStepStatus = "running" | "completed" | "error" | "waiting";
-
-interface ChatTraceStep {
-  id: string;
-  kind: "phase" | "tool";
-  title: string;
-  status: ChatTraceStepStatus;
-  details: string[];
-  order: number;
-}
-
-type ChatExecutionTimelineItem =
-  | {
-      id: string;
-      kind: "thinking";
-      text: string;
-      order: number;
-    }
-  | {
-      id: string;
-      kind: "trace_step";
-      step: ChatTraceStep;
-      order: number;
-    };
-
-interface PendingLocalAttachmentFile {
-  id: string;
-  source: "local-file";
-  file: File;
-}
-
-interface PendingExplorerAttachmentFile {
-  id: string;
-  source: "explorer-path";
-  absolutePath: string;
-  name: string;
-  mime_type?: string | null;
-  size_bytes: number;
-  kind: "image" | "file" | "folder";
-}
-
-type PendingAttachment =
-  | PendingLocalAttachmentFile
-  | PendingExplorerAttachmentFile;
-
-interface AttachmentListItem {
-  id: string;
-  kind: "image" | "file" | "folder";
-  name: string;
-  size_bytes: number;
-  workspace_path?: string;
-  file?: File;
-}
-
-interface ImageAttachmentPreviewState {
-  attachment: AttachmentListItem;
-  browserSnapshot: BrowserVisibleSnapshotPayload | null;
-  dataUrl: string;
-  isLoading: boolean;
-  errorMessage: string;
-}
-
-function attachmentLooksLikeImage(
-  name: string,
-  mimeType?: string | null,
-): boolean {
-  const normalizedMimeType = (mimeType ?? "").trim().toLowerCase();
-  if (normalizedMimeType.startsWith("image/")) {
-    return true;
-  }
-  return /\.(avif|bmp|gif|heic|heif|ico|jpe?g|png|svg|webp)$/i.test(
-    name.trim(),
-  );
-}
-
-function pendingAttachmentIsImage(attachment: PendingAttachment): boolean {
-  if (attachment.source === "local-file") {
-    return attachmentLooksLikeImage(attachment.file.name, attachment.file.type);
-  }
-  return (
-    attachment.kind === "image" ||
-    attachmentLooksLikeImage(attachment.name, attachment.mime_type)
-  );
-}
-
-function supportsImageInput(
-  inputModalities?: readonly string[] | null,
-): boolean {
-  if (!Array.isArray(inputModalities) || inputModalities.length === 0) {
-    return true;
-  }
-  return inputModalities.includes("image");
-}
-
-function imageInputUnsupportedMessage(modelLabel: string): string {
-  const normalizedModelLabel = modelLabel.trim();
-  if (!normalizedModelLabel) {
-    return "The selected model doesn't support image inputs.";
-  }
-  return `${normalizedModelLabel} doesn't support image inputs.`;
-}
-
-interface ChatModelOption {
-  value: string;
-  label: string;
-  selectedLabel?: string;
-  searchText?: string;
-  disabled?: boolean;
-  statusLabel?: string;
-}
-
-interface ChatModelOptionGroup {
-  label: string;
-  options: ChatModelOption[];
-}
-
-interface ChatComposerSlashCommandOption {
-  key: string;
-  kind: "skill";
-  command: string;
-  label: string;
-  description: string;
-  searchText: string;
-  skillId: string;
-}
-
-interface ChatComposerQuotedSkillItem {
-  skillId: string;
-  title: string;
-}
-
-interface StreamTelemetryEntry {
-  id: string;
-  at: string;
-  streamId: string;
-  transportType: string;
-  eventName: string;
-  eventType: string;
-  inputId: string;
-  sessionId: string;
-  action: string;
-  detail: string;
-}
-
-export type ArtifactBrowserFilter =
-  | "all"
-  | "documents"
-  | "images"
-  | "code"
-  | "links"
-  | "apps";
-
-const STREAM_ATTACH_PENDING = "__stream_attach_pending__";
-const STREAM_TELEMETRY_LIMIT = 240;
-const TOOL_TRACE_TERMINAL_PHASES = new Set(["completed", "failed", "error"]);
-const CHAT_AUTO_SCROLL_THRESHOLD_PX = 72;
-// Drives both the initial session-open fetch and each "load earlier" pull.
-// Was 10 originally — small enough that scroll-restoration after a prepend
-// often left the user still inside the 96px top threshold, immediately
-// triggering the next load. The runtime caps `limit` at 1000 (default 200);
-// 50 keeps the per-call work bounded while making any single load earn
-// enough vertical content (~25 turns) to push the user well past the
-// re-trigger threshold.
-const CHAT_HISTORY_PAGE_SIZE = 50;
-const CHAT_HISTORY_TOP_LOAD_THRESHOLD_PX = 96;
-const COMPOSER_FOOTER_GAP_PX = 8;
-const COMPOSER_FULL_MODEL_CONTROL_WIDTH_PX = 240;
-const COMPOSER_FULL_THINKING_CONTROL_WIDTH_PX = 88;
-const COMPOSER_FULL_PROVIDER_SETUP_WIDTH_PX = 320;
-const COMPOSER_COMPACT_MODEL_CONTROL_MAX_WIDTH_PX = 168;
-const COMPOSER_COMPACT_THINKING_CONTROL_MIN_WIDTH_PX = 56;
-const COMPOSER_COMPACT_THINKING_CONTROL_MAX_WIDTH_PX = 124;
-const CHAT_MODEL_STORAGE_KEY = "holaboss-chat-model-v1";
-const CHAT_THINKING_STORAGE_KEY = "holaboss-chat-thinking-v1";
-const CHAT_MODEL_USE_RUNTIME_DEFAULT = "__runtime_default__";
-const CHAT_SERIALIZED_SKILL_COMMAND_PATTERN = /^\/([A-Za-z0-9_-]+)$/;
-const QUEUED_MESSAGES_PREVIEW_EVENT = "holaboss:queued-messages-preview-change";
-const LEGACY_UNAVAILABLE_CHAT_MODELS = new Set(["openai/gpt-5.2-mini"]);
-const DEPRECATED_CHAT_MODELS = new Set([
-  "openai/gpt-5.1",
-  "openai/gpt-5.1-codex",
-  "openai/gpt-5.1-codex-mini",
-  "openai/gpt-5.1-codex-max",
-  "gpt-5.1",
-  "gpt-5.1-codex",
-  "gpt-5.1-codex-mini",
-  "gpt-5.1-codex-max",
-]);
-const CHAT_MODEL_PRESETS = [
-  "openai/gpt-5.1",
-  "openai/gpt-5",
-  "openai/gpt-5.2",
-] as const;
-const RUNTIME_MODEL_CAPABILITY_ALIASES: Record<string, string> = {
-  chat: "chat",
-  text: "chat",
-  completion: "chat",
-  completions: "chat",
-  responses: "chat",
-  image: "image_generation",
-  images: "image_generation",
-  image_generation: "image_generation",
-  image_gen: "image_generation",
+export type {
+  ChatAssistantSegment,
+  ChatMessage,
+  ChatComposerMentionItem,
+  ArtifactBrowserFilter,
+};
+export {
+  inputIdFromMessageId,
+  historyMessagesInDisplayOrder,
+  turnInputIdsFromHistoryMessages,
+  ArtifactBrowserModal,
+  sortOutputs,
+  AssistantTurn,
+  UserTurn,
+  ConversationTurns,
 };
 
 function sessionUserId(
@@ -606,283 +419,6 @@ function serializeQuotedSkillPrompt(
     return lines.join("\n");
   }
   return [...lines, "", normalizedBody].join("\n");
-}
-
-function parseSerializedQuotedSkillPrompt(
-  value: string,
-): ChatSerializedQuotedSkillBlock {
-  const normalized = value.replace(/\r\n/g, "\n");
-  const lines = normalized.split("\n");
-  const skillIds: string[] = [];
-  let index = 0;
-
-  while (index < lines.length) {
-    const line = lines[index]?.trim() ?? "";
-    if (!line) {
-      break;
-    }
-    const match = CHAT_SERIALIZED_SKILL_COMMAND_PATTERN.exec(line);
-    if (!match) {
-      return {
-        skillIds: [],
-        body: normalized.trim(),
-      };
-    }
-    skillIds.push(match[1] ?? "");
-    index += 1;
-  }
-
-  if (skillIds.length === 0) {
-    return {
-      skillIds: [],
-      body: normalized.trim(),
-    };
-  }
-
-  if (index < lines.length && (lines[index]?.trim() ?? "") !== "") {
-    return {
-      skillIds: [],
-      body: normalized.trim(),
-    };
-  }
-
-  while (index < lines.length && (lines[index]?.trim() ?? "") === "") {
-    index += 1;
-  }
-
-  return {
-    skillIds: [...new Set(skillIds)],
-    body: lines.slice(index).join("\n").trim(),
-  };
-}
-
-function appendComposerPrefillText(currentInput: string, text: string) {
-  const normalizedText = text.trim();
-  if (!normalizedText) {
-    return currentInput;
-  }
-  if (!currentInput.trim()) {
-    return normalizedText;
-  }
-  return /[\s(]$/.test(currentInput)
-    ? `${currentInput}${normalizedText}`
-    : `${currentInput} ${normalizedText}`;
-}
-
-function buildComposerSlashCommandOptions(
-  skills: WorkspaceSkillRecordPayload[],
-): ChatComposerSlashCommandOption[] {
-  return skills
-    .map((skill) => ({
-      key: `skill:${skill.skill_id}`,
-      kind: "skill" as const,
-      command: `/${skill.skill_id}`,
-      label: skill.title,
-      description: skill.summary,
-      searchText:
-        `${skill.skill_id} ${skill.title} ${skill.summary}`.toLowerCase(),
-      skillId: skill.skill_id,
-    }))
-    .sort((left, right) => left.command.localeCompare(right.command));
-}
-
-function findActiveSlashCommandRange(
-  value: string,
-  caretIndex: number,
-): { start: number; end: number; query: string } | null {
-  if (caretIndex < 0 || caretIndex > value.length) {
-    return null;
-  }
-  const prefix = value.slice(0, caretIndex);
-  const whitespaceBoundary = Math.max(
-    prefix.lastIndexOf(" "),
-    prefix.lastIndexOf("\n"),
-    prefix.lastIndexOf("\t"),
-  );
-  const start = whitespaceBoundary + 1;
-  const rawToken = prefix.slice(start);
-  if (!rawToken.startsWith("/") || rawToken.length === 0) {
-    return null;
-  }
-  if (!/^\/[A-Za-z0-9_-]*$/.test(rawToken)) {
-    return null;
-  }
-  return {
-    start,
-    end: caretIndex,
-    query: rawToken.slice(1).toLowerCase(),
-  };
-}
-
-function removeSlashCommandText(
-  value: string,
-  range: { start: number; end: number },
-): { value: string; caretIndex: number } {
-  const before = value.slice(0, range.start);
-  const after = value.slice(range.end);
-  const nextValue =
-    before.endsWith(" ") && after.startsWith(" ")
-      ? `${before}${after.slice(1)}`
-      : `${before}${after}`;
-  return {
-    value: nextValue,
-    caretIndex: before.length,
-  };
-}
-
-/** Mirrors `findActiveSlashCommandRange` for the `@`-mention trigger.
- *  Treated as a separate path so the two pickers can fire at different
- *  carets and against different item lists without one swallowing the
- *  other. The matched character set is wider than slash (allows `.`)
- *  so handles like `agent.work` resolve. */
-function findActiveMentionRange(
-  value: string,
-  caretIndex: number,
-): { start: number; end: number; query: string } | null {
-  if (caretIndex < 0 || caretIndex > value.length) {
-    return null;
-  }
-  const prefix = value.slice(0, caretIndex);
-  const whitespaceBoundary = Math.max(
-    prefix.lastIndexOf(" "),
-    prefix.lastIndexOf("\n"),
-    prefix.lastIndexOf("\t"),
-  );
-  const start = whitespaceBoundary + 1;
-  const rawToken = prefix.slice(start);
-  if (!rawToken.startsWith("@") || rawToken.length === 0) {
-    return null;
-  }
-  // Allow `/` in handles so nested file paths (`@apps/twitter/post.md`)
-  // round-trip — the parser still recognises the token if the user
-  // moves the caret back into it.
-  if (!/^@[A-Za-z0-9_.\-/]*$/.test(rawToken)) {
-    return null;
-  }
-  return {
-    start,
-    end: caretIndex,
-    query: rawToken.slice(1).toLowerCase(),
-  };
-}
-
-function replaceMentionText(
-  value: string,
-  range: { start: number; end: number },
-  insertion: string,
-): { value: string; caretIndex: number } {
-  const before = value.slice(0, range.start);
-  const after = value.slice(range.end);
-  // Trailing space follows the inserted handle so the user can keep
-  // typing without manually adding one. Mirrors what most chat apps
-  // do after a successful mention selection.
-  const trailing = after.startsWith(" ") || after.length === 0 ? "" : " ";
-  const nextValue = `${before}${insertion}${trailing}${after}`;
-  return {
-    value: nextValue,
-    caretIndex: before.length + insertion.length + trailing.length,
-  };
-}
-
-/** Token shape for `@`-mentions inside body text. Mirrors the rules
- *  in `findActiveMentionRange`: handle is `[A-Za-z0-9_.\-/]+`,
- *  preceded by start-of-string or whitespace. Backtick fences and
- *  inline-code spans are not yet skipped — a `@token` inside a
- *  ``` ``` ``` fence will still get rewritten. Acceptable for v1
- *  since user-submitted code is rare in chat. */
-const MENTION_TOKEN_PATTERN = /(^|[\s])@([A-Za-z0-9_.\-/]+)/g;
-
-/** Pre-process raw chat text so that `@<handle>` tokens become
- *  markdown links pointing at the `holaboss-mention://` scheme.
- *  SimpleMarkdown's link renderer (with `renderMention` configured)
- *  swaps each one for an inline `EntityMention` chip. Keeps markdown
- *  rendering otherwise untouched. */
-function injectMentionLinks(text: string): string {
-  if (!text.includes("@")) return text;
-  return text.replace(MENTION_TOKEN_PATTERN, (_match, leading, handle) => {
-    return `${leading}[@${handle}](holaboss-mention://${handle})`;
-  });
-}
-
-function displayModelLabel(model: string) {
-  const trimmed = model.trim();
-  if (!trimmed) {
-    return "Unknown model";
-  }
-
-  const withoutProvider = trimmed.replace(/^(openai|anthropic)\//i, "");
-  const sonnetModelMatch = withoutProvider.match(
-    /^claude-sonnet-(\d+)-(\d+)$/i,
-  );
-  if (sonnetModelMatch) {
-    return `Claude Sonnet ${sonnetModelMatch[1]}.${sonnetModelMatch[2]}`;
-  }
-
-  if (/^gpt-/i.test(withoutProvider)) {
-    return withoutProvider
-      .replace(/^gpt-/i, "GPT-")
-      .replace(/-mini\b/gi, " Mini")
-      .replace(/-codex\b/gi, " Codex")
-      .replace(/-max\b/gi, " Max")
-      .replace(/-spark\b/gi, " Spark");
-  }
-
-  return withoutProvider
-    .split(/[-_]/)
-    .filter(Boolean)
-    .map((part) =>
-      /^\d+(\.\d+)?$/.test(part)
-        ? part
-        : `${part[0]?.toUpperCase() ?? ""}${part.slice(1)}`,
-    )
-    .join(" ");
-}
-
-function compactComposerModelLabel(label: string) {
-  const normalizedLabel = label.trim();
-  if (!normalizedLabel) {
-    return "Model";
-  }
-
-  const autoMatch = normalizedLabel.match(/^Auto \((.+)\)$/i);
-  if (autoMatch?.[1]) {
-    return autoMatch[1].trim();
-  }
-
-  const segments = normalizedLabel
-    .split("·")
-    .map((segment) => segment.trim())
-    .filter(Boolean);
-  return segments[segments.length - 1] ?? normalizedLabel;
-}
-
-function displayThinkingValueLabel(value: string) {
-  const normalizedValue = value.trim().toLowerCase();
-  if (!normalizedValue) {
-    return "Thinking";
-  }
-
-  if (normalizedValue === "xhigh") {
-    return "Extra High";
-  }
-  if (
-    normalizedValue === "none" ||
-    normalizedValue === "minimal" ||
-    normalizedValue === "low" ||
-    normalizedValue === "medium" ||
-    normalizedValue === "high" ||
-    normalizedValue === "max"
-  ) {
-    return `${normalizedValue[0]?.toUpperCase() ?? ""}${normalizedValue.slice(1)}`;
-  }
-  if (/^-?\d+$/.test(normalizedValue)) {
-    return Number(normalizedValue).toLocaleString();
-  }
-  return normalizedValue
-    .split(/[-_]/)
-    .filter(Boolean)
-    .map((part) => `${part[0]?.toUpperCase() ?? ""}${part.slice(1)}`)
-    .join(" ");
 }
 
 function runtimeModelDisplayLabel(model: RuntimeProviderModelPayload) {
@@ -1255,225 +791,6 @@ export function chatMessagesFromSessionState(params: {
     );
 }
 
-function formatAttachmentSize(sizeBytes: number) {
-  if (!Number.isFinite(sizeBytes) || sizeBytes <= 0) {
-    return "";
-  }
-  if (sizeBytes >= 1024 * 1024) {
-    return `${(sizeBytes / (1024 * 1024)).toFixed(1)} MB`;
-  }
-  if (sizeBytes >= 1024) {
-    return `${Math.round(sizeBytes / 1024)} KB`;
-  }
-  return `${sizeBytes} B`;
-}
-
-function outputMetadataString(
-  output: WorkspaceOutputRecordPayload,
-  key: string,
-) {
-  const value = output.metadata[key];
-  return typeof value === "string" ? value.trim() : "";
-}
-
-function outputMetadataNumber(
-  output: WorkspaceOutputRecordPayload,
-  key: string,
-) {
-  const value = output.metadata[key];
-  return typeof value === "number" && Number.isFinite(value) ? value : null;
-}
-
-function outputBrowserFilterForOutput(
-  output: WorkspaceOutputRecordPayload,
-): ArtifactBrowserFilter {
-  if (
-    outputMetadataString(output, "origin_type") === "app" ||
-    output.module_id
-  ) {
-    return "apps";
-  }
-  const category = outputMetadataString(output, "category");
-  if (category === "image") {
-    return "images";
-  }
-  if (category === "code") {
-    return "code";
-  }
-  if (category === "link") {
-    return "links";
-  }
-  return "documents";
-}
-
-function outputKindLabel(output: WorkspaceOutputRecordPayload) {
-  // .dashboard files outrank generic origin/category labels — they're
-  // a first-class artifact type with their own renderer.
-  const ext = outputFileExtensionFromTitle(output);
-  if (ext === "dashboard") {
-    return "Dashboard";
-  }
-  if (
-    outputMetadataString(output, "origin_type") === "app" ||
-    output.module_id
-  ) {
-    const artifactType = outputMetadataString(output, "artifact_type");
-    if (artifactType) {
-      return artifactType.charAt(0).toUpperCase() + artifactType.slice(1);
-    }
-    return "Artifact";
-  }
-  const category = outputMetadataString(output, "category");
-  if (category === "image") {
-    return "Image";
-  }
-  if (category === "code") {
-    return "Code file";
-  }
-  if (category === "link") {
-    return "Link";
-  }
-  if (category === "spreadsheet") {
-    return "Spreadsheet";
-  }
-  if (category === "document") {
-    return "Document";
-  }
-  return output.output_type === "document" ? "Document" : "File";
-}
-
-// Lightweight extension lookup that doesn't depend on the metadata
-// envelope (which agent-authored files often skip). Reads the title
-// suffix only — used inside outputKindLabel before metadata parsing.
-function outputFileExtensionFromTitle(
-  output: WorkspaceOutputRecordPayload,
-): string {
-  const fromTitle = output.title?.trim() ?? "";
-  const dotIndex = fromTitle.lastIndexOf(".");
-  if (dotIndex > 0 && dotIndex < fromTitle.length - 1) {
-    return fromTitle.slice(dotIndex + 1).toLowerCase();
-  }
-  const path = outputMetadataString(output, "file_path") ?? output.file_path ?? "";
-  const pathDot = path.lastIndexOf(".");
-  if (pathDot > 0 && pathDot < path.length - 1) {
-    return path.slice(pathDot + 1).toLowerCase();
-  }
-  return "";
-}
-
-function outputChangeLabel(output: WorkspaceOutputRecordPayload) {
-  const changeType = outputMetadataString(output, "change_type");
-  if (changeType === "created") {
-    return "Created";
-  }
-  if (changeType === "modified") {
-    return "Updated";
-  }
-  return "";
-}
-
-function outputSecondaryLabel(output: WorkspaceOutputRecordPayload) {
-  const parts = [outputKindLabel(output)];
-  const sizeLabel = formatAttachmentSize(
-    outputMetadataNumber(output, "size_bytes") ?? 0,
-  );
-  if (sizeLabel) {
-    parts.push(sizeLabel);
-  }
-  const timeLabel = chatMessageTimeLabel(output.created_at);
-  if (timeLabel) {
-    parts.push(timeLabel);
-  }
-  return parts.join(" · ");
-}
-
-export function sortOutputs(outputs: WorkspaceOutputRecordPayload[]) {
-  return [...dedupeOutputsForDisplay(outputs)].sort((left, right) => {
-    const leftTime = Date.parse(left.created_at || "") || 0;
-    const rightTime = Date.parse(right.created_at || "") || 0;
-    if (leftTime !== rightTime) {
-      return leftTime - rightTime;
-    }
-    return left.title.localeCompare(right.title);
-  });
-}
-
-function sortOutputsLatestFirst(outputs: WorkspaceOutputRecordPayload[]) {
-  return [...dedupeOutputsForDisplay(outputs)].sort((left, right) => {
-    const leftTime = Date.parse(left.created_at || "") || 0;
-    const rightTime = Date.parse(right.created_at || "") || 0;
-    if (leftTime !== rightTime) {
-      return rightTime - leftTime;
-    }
-    return left.title.localeCompare(right.title);
-  });
-}
-
-function outputDisplayDedupeKey(output: WorkspaceOutputRecordPayload) {
-  const filePath = output.file_path?.trim() ?? "";
-  if (filePath) {
-    return `path:${filePath}`;
-  }
-  const artifactId = output.artifact_id?.trim() ?? "";
-  if (artifactId) {
-    return `artifact:${artifactId}`;
-  }
-  const title = output.title?.trim().toLowerCase() ?? "";
-  if (title) {
-    return `title:${title}`;
-  }
-  return `id:${output.id}`;
-}
-
-function outputDisplayPriority(output: WorkspaceOutputRecordPayload) {
-  let score = 0;
-  const originType = outputMetadataString(output, "origin_type");
-  if (originType === "forwarded_subagent") {
-    score += 40;
-  } else if (originType === "runtime_tool") {
-    score += 35;
-  } else if (originType === "app") {
-    score += 30;
-  }
-
-  if (outputMetadataString(output, "artifact_type") === "report") {
-    score += 20;
-  }
-  if (!/\.[A-Za-z0-9]+$/.test(output.title?.trim() ?? "")) {
-    score += 5;
-  }
-  return score;
-}
-
-function shouldPreferOutputForDisplay(
-  candidate: WorkspaceOutputRecordPayload,
-  current: WorkspaceOutputRecordPayload,
-) {
-  const candidatePriority = outputDisplayPriority(candidate);
-  const currentPriority = outputDisplayPriority(current);
-  if (candidatePriority !== currentPriority) {
-    return candidatePriority > currentPriority;
-  }
-  const candidateCreatedAt = Date.parse(candidate.created_at || "") || 0;
-  const currentCreatedAt = Date.parse(current.created_at || "") || 0;
-  if (candidateCreatedAt !== currentCreatedAt) {
-    return candidateCreatedAt > currentCreatedAt;
-  }
-  return candidate.title.localeCompare(current.title) < 0;
-}
-
-function dedupeOutputsForDisplay(outputs: WorkspaceOutputRecordPayload[]) {
-  const preferredByKey = new Map<string, WorkspaceOutputRecordPayload>();
-  for (const output of outputs) {
-    const key = outputDisplayDedupeKey(output);
-    const current = preferredByKey.get(key);
-    if (!current || shouldPreferOutputForDisplay(output, current)) {
-      preferredByKey.set(key, output);
-    }
-  }
-  return [...preferredByKey.values()];
-}
-
 function sortMemoryUpdateProposals(
   proposals: MemoryUpdateProposalRecordPayload[],
 ) {
@@ -1485,14 +802,6 @@ function sortMemoryUpdateProposals(
     }
     return left.title.localeCompare(right.title);
   });
-}
-
-function attachmentButtonLabel(attachment: {
-  name: string;
-  size_bytes: number;
-}) {
-  const sizeLabel = formatAttachmentSize(attachment.size_bytes);
-  return sizeLabel ? `${attachment.name} (${sizeLabel})` : attachment.name;
 }
 
 function attachmentUploadPayload(
@@ -1513,109 +822,6 @@ function attachmentUploadPayload(
     };
     reader.readAsDataURL(file);
   });
-}
-
-function attachmentFileExtension(mimeType?: string | null): string {
-  const normalizedMimeType = (mimeType ?? "").trim().toLowerCase();
-  if (!normalizedMimeType.includes("/")) {
-    return "bin";
-  }
-  const subtype = normalizedMimeType.split("/")[1]?.split("+")[0]?.trim() || "";
-  if (!subtype) {
-    return "bin";
-  }
-  if (subtype === "jpeg") {
-    return "jpg";
-  }
-  if (subtype === "svg") {
-    return "svg";
-  }
-  return subtype;
-}
-
-function normalizeClipboardAttachmentFile(file: File, index: number): File {
-  if (file.name.trim()) {
-    return file;
-  }
-
-  const extension = attachmentFileExtension(file.type);
-  const baseName = file.type.startsWith("image/")
-    ? `pasted-image-${index + 1}`
-    : `pasted-file-${index + 1}`;
-  return new File([file], `${baseName}.${extension}`, {
-    type: file.type,
-    lastModified: file.lastModified || Date.now(),
-  });
-}
-
-function clipboardFilesFromDataTransfer(
-  dataTransfer: DataTransfer | null,
-): File[] {
-  if (!dataTransfer) {
-    return [];
-  }
-
-  const clipboardFiles =
-    dataTransfer.files.length > 0
-      ? Array.from(dataTransfer.files)
-      : Array.from(dataTransfer.items ?? []).flatMap((item) => {
-          if (item.kind !== "file") {
-            return [];
-          }
-          const file = item.getAsFile();
-          return file ? [file] : [];
-        });
-
-  return clipboardFiles.map((file, index) =>
-    normalizeClipboardAttachmentFile(file, index),
-  );
-}
-
-function base64ToArrayBuffer(value: string): ArrayBuffer {
-  const binary = atob(value);
-  const bytes = new Uint8Array(binary.length);
-  for (let index = 0; index < binary.length; index += 1) {
-    bytes[index] = binary.charCodeAt(index);
-  }
-  return bytes.buffer;
-}
-
-function fileFromClipboardImagePayload(
-  payload: ClipboardImagePayload | null,
-): File | null {
-  const contentBase64 = payload?.content_base64?.trim() ?? "";
-  if (!payload || !contentBase64) {
-    return null;
-  }
-
-  try {
-    return new File([base64ToArrayBuffer(contentBase64)], payload.name, {
-      type: payload.mime_type || "image/png",
-      lastModified: Date.now(),
-    });
-  } catch {
-    return null;
-  }
-}
-
-async function clipboardImageFileFromElectronClipboard(): Promise<File | null> {
-  const payload = await window.electronAPI.clipboard.readImage();
-  return fileFromClipboardImagePayload(payload);
-}
-
-function explorerAttachmentFilesFromClipboardText(
-  clipboardText: string,
-): ExplorerAttachmentDragPayload[] {
-  const entry = getExplorerAttachmentClipboardEntry();
-  if (!entry) {
-    return [];
-  }
-
-  if (clipboardText.trim() !== entry.text) {
-    return [];
-  }
-
-  return [entry.payload];
 }
 
 function pendingAttachmentId(seed: string) {
@@ -1925,46 +1131,6 @@ function toolTraceStepId(payload: Record<string, unknown>) {
   return callId || toolId || toolName
     ? `tool:${callId || toolId || toolName}`
     : "";
-}
-
-export function inputIdFromMessageId(
-  messageId: string,
-  role: "user" | "assistant",
-) {
-  const prefix = `${role}-`;
-  return messageId.startsWith(prefix) ? messageId.slice(prefix.length) : "";
-}
-
-export function inputIdFromHistoryMessage(
-  message: SessionHistoryMessagePayload,
-) {
-  if (message.role === "user" || message.role === "assistant") {
-    return inputIdFromMessageId(message.id, message.role);
-  }
-  return "";
-}
-
-export function historyMessagesInDisplayOrder(
-  messages: SessionHistoryMessagePayload[],
-  order: "asc" | "desc",
-) {
-  return order === "desc" ? [...messages].reverse() : messages;
-}
-
-export function turnInputIdsFromHistoryMessages(
-  messages: SessionHistoryMessagePayload[],
-) {
-  const seen = new Set<string>();
-  const inputIds: string[] = [];
-  for (const message of messages) {
-    const inputId = inputIdFromHistoryMessage(message);
-    if (!inputId || seen.has(inputId)) {
-      continue;
-    }
-    seen.add(inputId);
-    inputIds.push(inputId);
-  }
-  return inputIds;
 }
 
 function assistantInputIdsFromChatMessages(messages: ChatMessage[]) {
@@ -2482,32 +1648,6 @@ function extractToolErrorText(payload: Record<string, unknown>): string {
   }
 
   return "";
-}
-
-function isIntegrationError(
-  text: string,
-): { provider: string; action: string } | null {
-  const patterns: Array<{ pattern: RegExp; provider: string }> = [
-    { pattern: /no\s+google\s+token/i, provider: "Google" },
-    { pattern: /no\s+github\s+token/i, provider: "GitHub" },
-    { pattern: /no\s+reddit\s+token/i, provider: "Reddit" },
-    { pattern: /no\s+twitter\s+token/i, provider: "Twitter" },
-    { pattern: /no\s+linkedin\s+token/i, provider: "LinkedIn" },
-    { pattern: /PLATFORM_INTEGRATION_TOKEN/i, provider: "" },
-    { pattern: /integration.*not.*connected/i, provider: "" },
-    { pattern: /integration.*not.*bound/i, provider: "" },
-    { pattern: /connect\s+via\s+(settings|integrations)/i, provider: "" },
-  ];
-  for (const { pattern, provider } of patterns) {
-    if (pattern.test(text)) {
-      const resolved = provider || "this provider";
-      return {
-        provider: resolved,
-        action: `Connect ${resolved} in the Integrations tab`,
-      };
-    }
-  }
-  return null;
 }
 
 function toolTraceStepFromPayload(
@@ -3243,45 +2383,6 @@ function latestVisibleChatMessageId(messages: ChatMessage[]): string {
     }
   }
   return "";
-}
-
-function chatMessageTimeLabel(value: string | null | undefined): string {
-  const timestamp = Date.parse(value || "");
-  if (Number.isNaN(timestamp)) {
-    return "";
-  }
-  return new Date(timestamp).toLocaleTimeString(undefined, {
-    hour: "numeric",
-    minute: "2-digit",
-  });
-}
-
-async function copyTextToClipboard(value: string) {
-  if (typeof navigator !== "undefined" && navigator.clipboard?.writeText) {
-    await navigator.clipboard.writeText(value);
-    return;
-  }
-  if (typeof document === "undefined") {
-    throw new Error("Clipboard is unavailable.");
-  }
-
-  const textarea = document.createElement("textarea");
-  textarea.value = value;
-  textarea.setAttribute("readonly", "");
-  textarea.style.position = "fixed";
-  textarea.style.opacity = "0";
-  textarea.style.pointerEvents = "none";
-  document.body.appendChild(textarea);
-  textarea.focus();
-  textarea.select();
-  textarea.setSelectionRange(0, textarea.value.length);
-
-  const copied = document.execCommand("copy");
-  document.body.removeChild(textarea);
-
-  if (!copied) {
-    throw new Error("Clipboard copy failed.");
-  }
 }
 
 function hasActiveChatSelection(container: HTMLDivElement | null) {
@@ -8365,3596 +7466,3 @@ export function ChatPane({
   );
 }
 
-interface ChatHeaderProps {
-  agentName: string;
-  workspace: WorkspaceRecordPayload | null;
-  subtitle?: string;
-  onReturnToMainSession?: () => void;
-  onOpenSessions?: () => void;
-  onOpenInbox?: () => void;
-  inboxUnreadCount: number;
-  onOpenAutomations?: () => void;
-}
-
-function ChatHeader({
-  agentName,
-  workspace,
-  subtitle,
-  onReturnToMainSession,
-  onOpenSessions,
-  onOpenInbox,
-  inboxUnreadCount,
-  onOpenAutomations,
-}: ChatHeaderProps) {
-  const seed = workspace?.id ?? agentName ?? "default";
-
-  return (
-    <div className="flex items-center justify-between gap-3">
-      <div className="flex min-w-0 flex-1 items-center gap-2">
-        <AgentAvatar seed={seed} size="sm" />
-        <div className="flex min-w-0 flex-col">
-          <span className="truncate text-sm font-medium text-foreground">
-            {agentName}
-          </span>
-          {subtitle ? (
-            <span className="truncate text-[11px] leading-tight text-muted-foreground">
-              {subtitle}
-            </span>
-          ) : null}
-        </div>
-      </div>
-
-      <div className="flex shrink-0 items-center gap-0.5">
-        {onOpenSessions ? (
-          <Button
-            type="button"
-            variant="ghost"
-            size="icon-sm"
-            onClick={() => onOpenSessions()}
-            aria-label="Sessions"
-            className="text-muted-foreground hover:text-foreground"
-          >
-            <MessageCircle className="size-4" />
-          </Button>
-        ) : null}
-
-        {onOpenInbox ? (
-          <Button
-            type="button"
-            variant="ghost"
-            size="icon-sm"
-            onClick={() => onOpenInbox()}
-            aria-label="Inbox"
-            className="relative text-muted-foreground hover:text-foreground"
-          >
-            <Inbox className="size-4" />
-            {inboxUnreadCount > 0 ? (
-              <StatusDot
-                variant="destructive"
-                size="sm"
-                className="absolute right-1 top-1 border border-card"
-              />
-            ) : null}
-          </Button>
-        ) : null}
-
-        {onOpenAutomations ? (
-          <Button
-            type="button"
-            variant="ghost"
-            size="icon-sm"
-            onClick={() => onOpenAutomations()}
-            aria-label="Automations"
-            className="text-muted-foreground hover:text-foreground"
-          >
-            <Clock3 className="size-4" />
-          </Button>
-        ) : null}
-      </div>
-    </div>
-  );
-}
-
-interface ComposerProps {
-  input: string;
-  quotedSkills: ChatComposerQuotedSkillItem[];
-  slashCommands: ChatComposerSlashCommandOption[];
-  attachments: AttachmentListItem[];
-  isResponding: boolean;
-  pausePending: boolean;
-  pauseDisabled: boolean;
-  disabled: boolean;
-  disabledReason?: string;
-  selectedModel: string;
-  resolvedModelLabel: string;
-  runtimeDefaultModelLabel: string;
-  modelOptions: ChatModelOption[];
-  modelOptionGroups: ChatModelOptionGroup[];
-  runtimeDefaultModelAvailable: boolean;
-  selectedThinkingValue: string | null;
-  thinkingValues: string[];
-  showThinkingValueSelector: boolean;
-  modelSelectionUnavailableReason: string;
-  submitDisabled?: boolean;
-  placeholder: string;
-  showModelSelector: boolean;
-  onModelChange: (value: string) => void;
-  onThinkingValueChange: (value: string | null) => void;
-  onOpenModelProviders: () => void;
-  textareaRef: RefObject<HTMLTextAreaElement | null>;
-  fileInputRef: RefObject<HTMLInputElement | null>;
-  onChange: (value: string) => void;
-  onKeyDown: (event: KeyboardEvent<HTMLTextAreaElement>) => void;
-  onCompositionStart: (event: CompositionEvent<HTMLTextAreaElement>) => void;
-  onCompositionEnd: (event: CompositionEvent<HTMLTextAreaElement>) => void;
-  onAttachmentInputChange: (event: ChangeEvent<HTMLInputElement>) => void;
-  onPause: () => void;
-  onAddDroppedFiles: (files: File[]) => void;
-  onAddExplorerAttachments: (files: ExplorerAttachmentDragPayload[]) => void;
-  onSelectSlashCommand: (command: ChatComposerSlashCommandOption) => void;
-  /** Items the `@` picker offers — currently workspaces, future:
-   *  apps / sessions / memories. Pre-shaped so the picker is just a
-   *  presenter; the parent decides what's mentionable. */
-  mentionableItems?: ChatComposerMentionItem[];
-  onRemoveQuotedSkill: (skillId: string) => void;
-  onRemoveAttachment: (attachmentId: string) => void;
-  onPreviewAttachment: (attachment: AttachmentListItem) => void;
-}
-
-export interface ChatComposerMentionItem {
-  id: string;
-  /** What gets inserted into the text — without the leading `@`. */
-  handle: string;
-  /** Visible label in the picker. Single-line; descriptions are
-   *  intentionally not part of this shape — quick pickers stay tight. */
-  label: ReactNode;
-  /** Tiny kind glyph (e.g. file/app icon) shown left of the label
-   *  so mixed-kind menus stay readable. */
-  kindIcon?: ReactNode;
-  /** Plain-text aliases for fuzzy match. */
-  keywords?: string[];
-}
-
-export const UserTurn = memo(UserTurnComponent, (prev, next) =>
-  prev.text === next.text &&
-  prev.createdAt === next.createdAt &&
-  prev.attachments === next.attachments,
-);
-
-function UserTurnComponent({
-  text,
-  createdAt,
-  attachments,
-  onPreviewAttachment,
-  onLinkClick,
-  onLocalLinkClick,
-}: {
-  text: string;
-  createdAt?: string;
-  attachments: ChatAttachment[];
-  onPreviewAttachment?: (attachment: AttachmentListItem) => void;
-  onLinkClick?: (url: string) => void;
-  onLocalLinkClick?: (href: string) => void;
-}) {
-  const [copyFeedbackVisible, setCopyFeedbackVisible] = useState(false);
-  const copyResetTimerRef = useRef<number | null>(null);
-  const timeLabel = chatMessageTimeLabel(createdAt);
-  const canCopy = text.trim().length > 0;
-  const showHoverFooter = canCopy || Boolean(timeLabel);
-  const parsedQuotedSkills = useMemo(
-    () => parseSerializedQuotedSkillPrompt(text),
-    [text],
-  );
-  const userBubbleText = parsedQuotedSkills.body || text.trim();
-
-  const bubbleContentRef = useRef<HTMLDivElement>(null);
-  const [isExpanded, setIsExpanded] = useState(false);
-  const [showExpandButton, setShowExpandButton] = useState(false);
-
-  useEffect(() => {
-    const node = bubbleContentRef.current;
-    if (!node) {
-      return;
-    }
-    // 180px ~= 6–7 lines of chat-user-markdown at 0.875rem / 1.6 leading.
-    setShowExpandButton(node.scrollHeight > 188);
-  }, [userBubbleText]);
-
-  useEffect(() => {
-    return () => {
-      if (copyResetTimerRef.current !== null) {
-        window.clearTimeout(copyResetTimerRef.current);
-      }
-    };
-  }, []);
-
-  const handleCopy = async () => {
-    if (!canCopy) {
-      return;
-    }
-
-    try {
-      await copyTextToClipboard(text);
-    } catch {
-      return;
-    }
-    setCopyFeedbackVisible(true);
-    if (copyResetTimerRef.current !== null) {
-      window.clearTimeout(copyResetTimerRef.current);
-    }
-    copyResetTimerRef.current = window.setTimeout(() => {
-      setCopyFeedbackVisible(false);
-      copyResetTimerRef.current = null;
-    }, 1600);
-  };
-
-  return (
-    <div className="group/user-turn flex min-w-0 justify-end">
-      <div
-        className={`relative z-0 flex min-w-0 max-w-[80%] flex-col items-end gap-2 group-hover/user-turn:z-10 group-focus-within/user-turn:z-10 ${showHoverFooter ? "pb-7" : ""}`.trim()}
-      >
-        {parsedQuotedSkills.skillIds.length > 0 ? (
-          <div className="flex max-w-full flex-wrap justify-end gap-1">
-            {parsedQuotedSkills.skillIds.map((skillId) => (
-              <EntityChip
-                key={skillId}
-                icon={<Sparkles className="text-muted-foreground" />}
-                label={`/${skillId}`}
-              />
-            ))}
-          </div>
-        ) : null}
-        {userBubbleText ? (
-          <div className="theme-chat-user-bubble inline-flex min-w-0 max-w-full flex-col items-stretch rounded-lg px-3 py-1.5 text-foreground">
-            <div
-              ref={bubbleContentRef}
-              className="relative overflow-hidden transition-[max-height] duration-300 ease-out"
-              style={{
-                maxHeight: showExpandButton && !isExpanded ? 180 : undefined,
-              }}
-            >
-              <SimpleMarkdown
-                className="chat-markdown chat-user-markdown max-w-full"
-                onLinkClick={onLinkClick}
-                onLocalLinkClick={onLocalLinkClick}
-                renderMention={(handle) => (
-                  <EntityMention label={handle} />
-                )}
-              >
-                {injectMentionLinks(userBubbleText)}
-              </SimpleMarkdown>
-              {showExpandButton && !isExpanded ? (
-                <div
-                  aria-hidden
-                  className="pointer-events-none absolute inset-x-0 bottom-0 h-10"
-                  style={{
-                    background:
-                      "linear-gradient(to bottom, transparent, color-mix(in oklch, var(--muted) 85%, var(--foreground) 4%))",
-                  }}
-                />
-              ) : null}
-            </div>
-            {showExpandButton ? (
-              <button
-                type="button"
-                onClick={() => setIsExpanded((value) => !value)}
-                className="mt-1.5 self-start text-xs font-medium text-muted-foreground transition-colors hover:text-foreground"
-              >
-                {isExpanded ? "Show less" : "Show more"}
-              </button>
-            ) : null}
-          </div>
-        ) : null}
-        {attachments.length > 0 ? (
-          <AttachmentList
-            attachments={attachments}
-            className="justify-end"
-            onPreview={onPreviewAttachment}
-          />
-        ) : null}
-        {showHoverFooter ? (
-          <div className="absolute bottom-0 right-1 flex w-max min-w-max max-w-none items-center gap-2 whitespace-nowrap text-xs text-muted-foreground opacity-0 pointer-events-none transition-opacity duration-150 group-hover/user-turn:opacity-100 group-hover/user-turn:pointer-events-auto group-focus-within/user-turn:opacity-100 group-focus-within/user-turn:pointer-events-auto">
-            {canCopy ? (
-              <Button
-                type="button"
-                variant="ghost"
-                size="icon-xs"
-                aria-label={
-                  copyFeedbackVisible
-                    ? "Copied user message"
-                    : "Copy user message"
-                }
-                onClick={() => {
-                  void handleCopy();
-                }}
-                className="size-6 rounded-lg text-muted-foreground hover:bg-fg-6 hover:text-foreground"
-              >
-                {copyFeedbackVisible ? (
-                  <Check className="size-3.5" strokeWidth={1.9} />
-                ) : (
-                  <Copy className="size-3.5" strokeWidth={1.9} />
-                )}
-              </Button>
-            ) : null}
-            {timeLabel ? (
-              <span className="select-none whitespace-nowrap tabular-nums">
-                {timeLabel}
-              </span>
-            ) : null}
-          </div>
-        ) : null}
-      </div>
-    </div>
-  );
-}
-
-function queuedSessionInputPreviewText(item: QueuedSessionInput) {
-  const parsedQuotedSkills = parseSerializedQuotedSkillPrompt(item.text);
-  const previewText =
-    parsedQuotedSkills.body ||
-    parsedQuotedSkills.skillIds.map((skillId) => `/${skillId}`).join(" ");
-  return previewText.replace(/\s+/g, " ").trim();
-}
-
-function QueuedSessionInputRail({
-  items,
-  onEditItem,
-  children,
-}: {
-  items: QueuedSessionInput[];
-  onEditItem?: (item: QueuedSessionInput, nextText: string) => Promise<void>;
-  children: ReactNode;
-}) {
-  const [editingInputId, setEditingInputId] = useState("");
-  const [editingDraft, setEditingDraft] = useState("");
-  const [editingError, setEditingError] = useState("");
-  const [savingInputId, setSavingInputId] = useState("");
-  const panelInsetPx = 18;
-  const panelHeightPx = 112;
-  const overlapPx = 28;
-  const queueViewportHeightPx = 50;
-  const reservedTopPx = 94;
-
-  useEffect(() => {
-    if (!editingInputId) {
-      return;
-    }
-    const activeItem = items.find((item) => item.inputId === editingInputId);
-    if (!activeItem || activeItem.status !== "queued") {
-      setEditingInputId("");
-      setEditingDraft("");
-      setEditingError("");
-      setSavingInputId("");
-    }
-  }, [editingInputId, items]);
-
-  const cancelEditing = () => {
-    setEditingInputId("");
-    setEditingDraft("");
-    setEditingError("");
-    setSavingInputId("");
-  };
-
-  const saveEditingItem = async (item: QueuedSessionInput) => {
-    if (!onEditItem || savingInputId || item.status !== "queued") {
-      return;
-    }
-    setEditingError("");
-    setSavingInputId(item.inputId);
-    try {
-      await onEditItem(item, editingDraft);
-      cancelEditing();
-    } catch (error) {
-      setEditingError(normalizeErrorMessage(error));
-    } finally {
-      setSavingInputId("");
-    }
-  };
-
-  if (items.length === 0) {
-    return <>{children}</>;
-  }
-
-  return (
-    <div className="relative" style={{ paddingTop: `${reservedTopPx}px` }}>
-      <div className="pointer-events-none absolute inset-x-0 top-0">
-        <div
-          className="pointer-events-auto absolute inset-x-0 overflow-hidden rounded-3xl bg-background shadow-md"
-          style={{
-            left: `${panelInsetPx}px`,
-            right: `${panelInsetPx}px`,
-            height: `${panelHeightPx}px`,
-          }}
-        >
-          <div className="px-4 pt-4">
-            <div
-              className="overflow-y-auto pr-1.5"
-              style={{ maxHeight: `${queueViewportHeightPx}px` }}
-            >
-              <div className="space-y-1.5">
-                {items.map((item) => {
-                  const previewText = queuedSessionInputPreviewText(item);
-                  const isEditing = editingInputId === item.inputId;
-                  const isSaving = savingInputId === item.inputId;
-                  return (
-                    <div
-                      key={item.inputId}
-                      className="rounded-xl px-1 text-sm leading-7 text-foreground"
-                    >
-                      {isEditing ? (
-                        <div className="space-y-1.5">
-                          <div className="flex items-center gap-2">
-                            <CornerDownLeft className="size-4 shrink-0 text-muted-foreground" />
-                            <Input
-                              value={editingDraft}
-                              onChange={(event) =>
-                                setEditingDraft(event.target.value)
-                              }
-                              onKeyDown={(event) => {
-                                if (event.key === "Enter") {
-                                  event.preventDefault();
-                                  void saveEditingItem(item);
-                                } else if (event.key === "Escape") {
-                                  event.preventDefault();
-                                  cancelEditing();
-                                }
-                              }}
-                              disabled={isSaving}
-                              autoFocus
-                              className="h-8 min-w-0 flex-1 rounded-lg border-border bg-background px-2.5 text-sm"
-                            />
-                            <Button
-                              type="button"
-                              variant="ghost"
-                              size="icon-xs"
-                              disabled={isSaving}
-                              onClick={() => {
-                                void saveEditingItem(item);
-                              }}
-                              className="size-7 rounded-full text-muted-foreground hover:bg-fg-6 hover:text-foreground"
-                              aria-label="Save queued message edit"
-                            >
-                              {isSaving ? (
-                                <Loader2 className="size-3.5 animate-spin" />
-                              ) : (
-                                <Check className="size-3.5" />
-                              )}
-                            </Button>
-                            <Button
-                              type="button"
-                              variant="ghost"
-                              size="icon-xs"
-                              disabled={isSaving}
-                              onClick={cancelEditing}
-                              className="size-7 rounded-full text-muted-foreground hover:bg-fg-6 hover:text-foreground"
-                              aria-label="Cancel queued message edit"
-                            >
-                              <X className="size-3.5" />
-                            </Button>
-                          </div>
-                          {editingError ? (
-                            <div className="pl-6 text-xs leading-5 text-destructive">
-                              {editingError}
-                            </div>
-                          ) : null}
-                        </div>
-                      ) : (
-                        <div className="flex items-center gap-3">
-                          <CornerDownLeft className="size-4 shrink-0 text-muted-foreground" />
-                          <div className="min-w-0 flex-1 truncate">
-                            {previewText || "Queued message"}
-                          </div>
-                          {onEditItem && item.status === "queued" ? (
-                            <Button
-                              type="button"
-                              variant="ghost"
-                              size="icon-xs"
-                              onClick={() => {
-                                setEditingInputId(item.inputId);
-                                setEditingDraft(previewText);
-                                setEditingError("");
-                              }}
-                              className="size-7 rounded-full text-muted-foreground hover:bg-fg-6 hover:text-foreground"
-                              aria-label="Edit queued message"
-                            >
-                              <PencilLine className="size-3.5" />
-                            </Button>
-                          ) : null}
-                        </div>
-                      )}
-                    </div>
-                  );
-                })}
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-      <div
-        className="relative z-10 rounded-3xl bg-background"
-        style={{ marginTop: `${-overlapPx}px` }}
-      >
-        {children}
-      </div>
-    </div>
-  );
-}
-
-function executionItemsHaveFileEdits(
-  items: ChatExecutionTimelineItem[],
-): boolean {
-  if (items.length === 0) {
-    return false;
-  }
-  return items.some((item) => {
-    if (item.kind !== "trace_step" || item.step.kind !== "tool") {
-      return false;
-    }
-    const title = item.step.title.toLowerCase();
-    return (
-      title.startsWith("edit") ||
-      title.startsWith("write") ||
-      title.startsWith("patch") ||
-      title.startsWith("replace") ||
-      title.startsWith("multiedit") ||
-      title.startsWith("apply") ||
-      title.startsWith("create file")
-    );
-  });
-}
-
-interface AssistantTurnActionsMenuProps {
-  copyText: string;
-  onViewTurnDetails?: () => void;
-  onViewFileChanges?: () => void;
-  hasFileEdits?: boolean;
-}
-
-function AssistantTurnActionsMenu({
-  copyText,
-  onViewTurnDetails,
-  onViewFileChanges,
-  hasFileEdits,
-}: AssistantTurnActionsMenuProps) {
-  const [copied, setCopied] = useState(false);
-
-  async function handleCopy() {
-    if (!copyText) {
-      return;
-    }
-    try {
-      await navigator.clipboard.writeText(copyText);
-      setCopied(true);
-      setTimeout(() => setCopied(false), 1500);
-    } catch {
-      // ignore
-    }
-  }
-
-  const showFileChanges = Boolean(onViewFileChanges) && Boolean(hasFileEdits);
-  const canCopy = copyText.trim().length > 0;
-  const hasOverflowItems =
-    Boolean(onViewTurnDetails) || showFileChanges;
-
-  if (!canCopy && !hasOverflowItems) {
-    return null;
-  }
-
-  return (
-    <div className="flex items-center gap-0.5">
-      {canCopy ? (
-        <Button
-          aria-label={copied ? "Copied message" : "Copy message"}
-          className="size-6 rounded-lg text-muted-foreground hover:bg-fg-6 hover:text-foreground"
-          size="icon-xs"
-          type="button"
-          variant="ghost"
-          onClick={() => void handleCopy()}
-        >
-          {copied ? (
-            <Check className="size-3.5" strokeWidth={1.9} />
-          ) : (
-            <Copy className="size-3.5" strokeWidth={1.9} />
-          )}
-        </Button>
-      ) : null}
-      {hasOverflowItems ? (
-        <DropdownMenu>
-          <DropdownMenuTrigger
-            render={
-              <Button
-                aria-label="Turn actions"
-                className="size-6 rounded-lg text-muted-foreground hover:bg-fg-6 hover:text-foreground"
-                size="icon-xs"
-                type="button"
-                variant="ghost"
-              >
-                <MoreHorizontal className="size-3.5" strokeWidth={1.9} />
-              </Button>
-            }
-          />
-          <DropdownMenuContent align="end" className="w-44" sideOffset={4}>
-            {onViewTurnDetails ? (
-              <DropdownMenuItem onClick={onViewTurnDetails}>
-                <ListTree />
-                View turn details
-              </DropdownMenuItem>
-            ) : null}
-            {showFileChanges ? (
-              <DropdownMenuItem onClick={onViewFileChanges}>
-                <FileCode2 />
-                View file changes
-              </DropdownMenuItem>
-            ) : null}
-          </DropdownMenuContent>
-        </DropdownMenu>
-      ) : null}
-    </div>
-  );
-}
-
-export const AssistantTurn = memo(AssistantTurnComponent, (prev, next) =>
-  prev.label === next.label &&
-  prev.mode === next.mode &&
-  prev.showExecutionInternals === next.showExecutionInternals &&
-  prev.text === next.text &&
-  prev.tone === next.tone &&
-  prev.segments === next.segments &&
-  prev.executionItems === next.executionItems &&
-  prev.memoryProposals === next.memoryProposals &&
-  prev.outputs === next.outputs &&
-  prev.memoryProposalAction === next.memoryProposalAction &&
-  prev.editingMemoryProposalId === next.editingMemoryProposalId &&
-  prev.memoryProposalDrafts === next.memoryProposalDrafts &&
-  prev.collapsedTraceByStepId === next.collapsedTraceByStepId &&
-  prev.live === next.live &&
-  prev.status === next.status &&
-  prev.statusAccessory === next.statusAccessory &&
-  prev.footerAccessory === next.footerAccessory,
-);
-
-function AssistantTurnComponent({
-  label,
-  mode,
-  showExecutionInternals = true,
-  fitToContent = false,
-  text,
-  tone = "default",
-  segments,
-  executionItems,
-  memoryProposals,
-  outputs,
-  memoryProposalAction,
-  editingMemoryProposalId,
-  memoryProposalDrafts,
-  onEditMemoryProposal,
-  onMemoryProposalDraftChange,
-  onAcceptMemoryProposal,
-  onDismissMemoryProposal,
-  onOpenOutput,
-  onOpenAllArtifacts,
-  collapsedTraceByStepId,
-  onToggleTraceStep,
-  onLinkClick,
-  onLocalLinkClick,
-  showAvatar = false,
-  workspaceId = null,
-  createdAt,
-  status = "",
-  live = false,
-  statusAccessory = null,
-  footerAccessory = null,
-}: {
-  label: string;
-  mode: string;
-  showExecutionInternals?: boolean;
-  fitToContent?: boolean;
-  text: string;
-  tone?: ChatMessage["tone"];
-  segments: ChatAssistantSegment[];
-  executionItems: ChatExecutionTimelineItem[];
-  memoryProposals: MemoryUpdateProposalRecordPayload[];
-  outputs: WorkspaceOutputRecordPayload[];
-  memoryProposalAction: {
-    proposalId: string;
-    action: "accept" | "dismiss";
-  } | null;
-  editingMemoryProposalId: string | null;
-  memoryProposalDrafts: Record<string, string>;
-  onEditMemoryProposal: (proposalId: string) => void;
-  onMemoryProposalDraftChange: (proposalId: string, value: string) => void;
-  onAcceptMemoryProposal: (proposal: MemoryUpdateProposalRecordPayload) => void;
-  onDismissMemoryProposal: (
-    proposal: MemoryUpdateProposalRecordPayload,
-  ) => void;
-  onOpenOutput?: (output: WorkspaceOutputRecordPayload) => void;
-  onOpenAllArtifacts: (outputs: WorkspaceOutputRecordPayload[]) => void;
-  collapsedTraceByStepId: Record<string, boolean>;
-  onToggleTraceStep: (stepId: string) => void;
-  onLinkClick?: (url: string) => void;
-  onLocalLinkClick?: (href: string) => void;
-  showAvatar?: boolean;
-  workspaceId?: string | null;
-  createdAt?: string;
-  status?: string;
-  live?: boolean;
-  statusAccessory?: ReactNode;
-  footerAccessory?: ReactNode;
-}) {
-  const normalizedStatus = (
-    showExecutionInternals ? status : status ? "Working" : ""
-  )
-    .replace(/\.+$/, "")
-    .trim();
-  const visibleSegments = showExecutionInternals
-    ? segments
-    : segments.filter(
-        (segment): segment is Extract<ChatAssistantSegment, { kind: "output" }> =>
-          segment.kind === "output",
-      );
-  const visibleExecutionItems = showExecutionInternals ? executionItems : [];
-  const renderedSegments =
-    visibleSegments.length > 0
-      ? visibleSegments
-      : visibleExecutionItems.length > 0 || Boolean(text)
-        ? [
-            ...(visibleExecutionItems.length > 0
-              ? ([
-                  {
-                    kind: "execution",
-                    items: visibleExecutionItems,
-                  },
-                ] as ChatAssistantSegment[])
-              : []),
-            ...(text
-              ? ([
-                  {
-                    kind: "output",
-                    text,
-                    tone,
-                  },
-                ] as ChatAssistantSegment[])
-              : []),
-          ]
-        : [];
-  const showStatusPlaceholder =
-    live && Boolean(normalizedStatus) && renderedSegments.length === 0;
-  const lastSegmentIsOutput =
-    renderedSegments.length > 0 &&
-    renderedSegments[renderedSegments.length - 1]?.kind === "output";
-  const showWorkingStatusLine =
-    live &&
-    showExecutionInternals &&
-    renderedSegments.length > 0 &&
-    !lastSegmentIsOutput;
-  const showStreamingCursor = live && lastSegmentIsOutput;
-
-  const [forceExpandToken, setForceExpandToken] = useState(0);
-  const hasFileEdits = useMemo(
-    () => executionItemsHaveFileEdits(executionItems),
-    [executionItems],
-  );
-  const copyText = useMemo(
-    () =>
-      renderedSegments
-        .filter(
-          (segment): segment is Extract<ChatAssistantSegment, { kind: "output" }> =>
-            segment.kind === "output",
-        )
-        .map((segment) => segment.text)
-        .join("\n\n")
-        .trim() || text.trim(),
-    [renderedSegments, text],
-  );
-  const hasAnyContent = renderedSegments.length > 0;
-  const showActionsMenu = hasAnyContent && !live;
-  const renderStatusLine = (nextLabel: string, className = "") => {
-    const resolvedLabel = nextLabel.trim() || "Working";
-    if (!statusAccessory) {
-      return <LiveStatusLine label={resolvedLabel} className={className} />;
-    }
-    return (
-      <div
-        className={`flex min-w-0 items-center justify-between gap-3 ${className}`.trim()}
-      >
-        <LiveStatusLine label={resolvedLabel} className="min-w-0" />
-        <div className="shrink-0">{statusAccessory}</div>
-      </div>
-    );
-  };
-
-  const timeLabel = chatMessageTimeLabel(createdAt);
-
-  return (
-    <div
-      className="group/assistant-turn relative flex min-w-0 flex-col items-start animate-in fade-in-0 slide-in-from-bottom-1 duration-200"
-    >
-      <div className="flex w-full min-w-0 items-end gap-2">
-        <div className="w-5 shrink-0">
-          {showAvatar && workspaceId ? (
-            <AgentAvatar seed={workspaceId} size="sm" />
-          ) : null}
-        </div>
-      <article
-        className={
-          fitToContent
-            ? "min-w-0 inline-flex w-fit max-w-full flex-col rounded-lg bg-fg-6 px-3 py-2"
-            : "min-w-0 w-full max-w-4xl rounded-lg bg-fg-6 px-3 py-2"
-        }
-      >
-        {showStatusPlaceholder ? renderStatusLine(normalizedStatus) : null}
-
-        {renderedSegments.map((segment, index) =>
-          segment.kind === "execution" ? (
-            <TraceStepGroup
-              key={`execution-${index}`}
-              items={segment.items}
-              collapsedByStepId={collapsedTraceByStepId}
-              onToggleStep={onToggleTraceStep}
-              live={live}
-              liveOutputStarted={
-                live &&
-                renderedSegments
-                  .slice(index + 1)
-                  .some((nextSegment) => nextSegment.kind === "output")
-              }
-              onLinkClick={onLinkClick}
-              onLocalLinkClick={onLocalLinkClick}
-              forceExpandToken={forceExpandToken}
-            />
-          ) : segment.tone === "error" ? (
-            <div
-              key={`output-${index}`}
-              className="theme-chat-system-bubble mt-2 first:mt-0 rounded-xl border px-3 py-2.5 text-xs text-foreground"
-            >
-              <div className="flex items-center gap-2">
-                <AlertTriangle className="size-3.5 shrink-0 text-destructive" />
-                <SimpleMarkdown
-                  className="chat-markdown max-w-full text-foreground"
-                  onLinkClick={onLinkClick}
-                  onLocalLinkClick={onLocalLinkClick}
-                >
-                  {segment.text}
-                </SimpleMarkdown>
-              </div>
-            </div>
-          ) : (
-            <SimpleMarkdown
-              key={`output-${index}`}
-              className="chat-markdown chat-assistant-markdown mt-2 first:mt-0 max-w-full text-foreground"
-              onLinkClick={onLinkClick}
-              onLocalLinkClick={onLocalLinkClick}
-            >
-              {segment.text}
-            </SimpleMarkdown>
-          ),
-        )}
-
-        {showWorkingStatusLine
-          ? renderStatusLine(
-              "Working",
-              renderedSegments.some((segment) => segment.kind === "execution")
-                ? ""
-                : "",
-            )
-          : null}
-
-        {showStreamingCursor ? <StreamingCursor /> : null}
-
-        {footerAccessory ? (
-          <div className="mt-2 flex justify-start">{footerAccessory}</div>
-        ) : null}
-
-        {memoryProposals.length > 0 ? (
-          <AssistantTurnMemoryProposals
-            proposals={memoryProposals}
-            proposalAction={memoryProposalAction}
-            editingProposalId={editingMemoryProposalId}
-            drafts={memoryProposalDrafts}
-            onEditProposal={onEditMemoryProposal}
-            onDraftChange={onMemoryProposalDraftChange}
-            onAcceptProposal={onAcceptMemoryProposal}
-            onDismissProposal={onDismissMemoryProposal}
-          />
-        ) : null}
-
-        {outputs.length > 0 ? (
-          <AssistantTurnOutputs
-            outputs={outputs}
-            onOpenOutput={onOpenOutput}
-            onOpenAllArtifacts={onOpenAllArtifacts}
-          />
-        ) : null}
-      </article>
-
-      </div>
-      {showActionsMenu || (showAvatar && timeLabel) ? (
-        <div className="flex h-6 items-center gap-2 pl-9">
-          {showAvatar && timeLabel ? (
-            <span className="select-none text-[10px] leading-none text-muted-foreground tabular-nums">
-              {timeLabel}
-            </span>
-          ) : null}
-          {showActionsMenu ? (
-            <div className="pointer-events-none opacity-0 transition-opacity duration-150 group-hover/assistant-turn:pointer-events-auto group-hover/assistant-turn:opacity-100 group-focus-within/assistant-turn:pointer-events-auto group-focus-within/assistant-turn:opacity-100">
-              <AssistantTurnActionsMenu
-                copyText={copyText}
-                hasFileEdits={hasFileEdits}
-                onViewFileChanges={
-                  hasFileEdits
-                    ? () => setForceExpandToken((token) => token + 1)
-                    : undefined
-                }
-                onViewTurnDetails={
-                  executionItems.length > 0
-                    ? () => setForceExpandToken((token) => token + 1)
-                    : undefined
-                }
-              />
-            </div>
-          ) : null}
-        </div>
-      ) : null}
-    </div>
-  );
-}
-
-type OutputVisualKind =
-  | "spreadsheet"
-  | "document"
-  | "pdf"
-  | "code"
-  | "image"
-  | "link"
-  | "app"
-  | "dashboard"
-  | "file";
-
-const DASHBOARD_EXTENSIONS = new Set(["dashboard"]);
-
-const SPREADSHEET_EXTENSIONS = new Set([
-  "xlsx",
-  "xls",
-  "xlsm",
-  "xlsb",
-  "ods",
-  "csv",
-  "tsv",
-]);
-const PDF_EXTENSIONS = new Set(["pdf"]);
-const CODE_EXTENSIONS = new Set([
-  "ts",
-  "tsx",
-  "js",
-  "jsx",
-  "py",
-  "rb",
-  "go",
-  "rs",
-  "java",
-  "c",
-  "cpp",
-  "h",
-  "cs",
-  "php",
-  "swift",
-  "kt",
-  "sh",
-  "json",
-  "yml",
-  "yaml",
-  "toml",
-  "xml",
-  "sql",
-  "css",
-  "scss",
-]);
-const DOCUMENT_EXTENSIONS = new Set([
-  "md",
-  "mdx",
-  "markdown",
-  "txt",
-  "doc",
-  "docx",
-  "rtf",
-  "odt",
-  "html",
-  "htm",
-]);
-
-function outputFileExtension(output: WorkspaceOutputRecordPayload): string {
-  const metadataExt = outputMetadataString(output, "extension");
-  if (metadataExt) {
-    return metadataExt.replace(/^\./, "").toLowerCase();
-  }
-  const fromTitle = output.title?.trim() ?? "";
-  const dotIndex = fromTitle.lastIndexOf(".");
-  if (dotIndex > 0 && dotIndex < fromTitle.length - 1) {
-    return fromTitle.slice(dotIndex + 1).toLowerCase();
-  }
-  return "";
-}
-
-function outputVisualKind(
-  output: WorkspaceOutputRecordPayload,
-): OutputVisualKind {
-  const filter = outputBrowserFilterForOutput(output);
-  if (filter === "apps") {
-    return "app";
-  }
-  if (filter === "images") {
-    return "image";
-  }
-  if (filter === "links") {
-    return "link";
-  }
-
-  const extension = outputFileExtension(output);
-  if (extension) {
-    if (DASHBOARD_EXTENSIONS.has(extension)) {
-      return "dashboard";
-    }
-    if (SPREADSHEET_EXTENSIONS.has(extension)) {
-      return "spreadsheet";
-    }
-    if (PDF_EXTENSIONS.has(extension)) {
-      return "pdf";
-    }
-    if (CODE_EXTENSIONS.has(extension)) {
-      return "code";
-    }
-    if (DOCUMENT_EXTENSIONS.has(extension)) {
-      return "document";
-    }
-  }
-
-  if (filter === "code") {
-    return "code";
-  }
-  const category = outputMetadataString(output, "category");
-  if (category === "spreadsheet") {
-    return "spreadsheet";
-  }
-  if (category === "document") {
-    return "document";
-  }
-  return "file";
-}
-
-function outputVisualTheme(kind: OutputVisualKind): {
-  Icon: typeof FileText;
-  tileClass: string;
-  iconClass: string;
-} {
-  switch (kind) {
-    case "spreadsheet":
-      return {
-        Icon: FileSpreadsheet,
-        tileClass: "bg-success/12 ring-1 ring-inset ring-success/20",
-        iconClass: "text-success",
-      };
-    case "pdf":
-      return {
-        Icon: FileType,
-        tileClass: "bg-destructive/12 ring-1 ring-inset ring-destructive/20",
-        iconClass: "text-destructive",
-      };
-    case "document":
-      return {
-        Icon: FileText,
-        tileClass: "bg-info/12 ring-1 ring-inset ring-info/20",
-        iconClass: "text-info",
-      };
-    case "code":
-      return {
-        Icon: FileCode2,
-        tileClass: "bg-info/12 ring-1 ring-inset ring-info/20",
-        iconClass: "text-info",
-      };
-    case "image":
-      return {
-        Icon: ImageIcon,
-        tileClass: "bg-warning/12 ring-1 ring-inset ring-warning/20",
-        iconClass: "text-warning",
-      };
-    case "link":
-      return {
-        Icon: Link2,
-        tileClass: "bg-info/12 ring-1 ring-inset ring-info/20",
-        iconClass: "text-info",
-      };
-    case "app":
-      return {
-        Icon: Waypoints,
-        tileClass: "bg-primary/12 ring-1 ring-inset ring-primary/20",
-        iconClass: "text-primary",
-      };
-    case "dashboard":
-      return {
-        Icon: LayoutDashboard,
-        tileClass: "bg-primary/12 ring-1 ring-inset ring-primary/20",
-        iconClass: "text-primary",
-      };
-    default:
-      return {
-        Icon: FileIcon,
-        tileClass: "bg-muted ring-1 ring-inset ring-border",
-        iconClass: "text-muted-foreground",
-      };
-  }
-}
-
-function OutputArtifactIcon({
-  output,
-  size = "md",
-}: {
-  output: WorkspaceOutputRecordPayload;
-  size?: "sm" | "md";
-}) {
-  const kind = outputVisualKind(output);
-  const { Icon, tileClass, iconClass } = outputVisualTheme(kind);
-  const tileSize = size === "sm" ? "size-7" : "size-9";
-  const iconSize = size === "sm" ? 14 : 16;
-  return (
-    <div
-      className={`grid ${tileSize} shrink-0 place-items-center rounded-lg ${tileClass}`}
-    >
-      <Icon size={iconSize} className={iconClass} />
-    </div>
-  );
-}
-
-function HistoryRestoreSkeleton() {
-  return (
-    <div
-      role="status"
-      aria-label="Loading conversation"
-      className="absolute inset-0 z-30 overflow-hidden px-6 pb-5 pt-5"
-    >
-      <div className="flex h-full flex-col">
-        <div className="animate-pulse space-y-6">
-          <div className="flex items-start justify-between gap-6">
-            <div className="h-5 w-28 rounded-md bg-muted" />
-            <div className="h-11 w-52 rounded-2xl bg-muted" />
-          </div>
-          <div className="space-y-3 px-3">
-            <div className="flex items-center gap-2">
-              <div className="h-5 w-6 rounded-md bg-muted" />
-              <div className="h-5 w-14 rounded-md bg-muted" />
-            </div>
-            <div className="h-5 w-full rounded-md bg-muted" />
-            <div className="h-5 w-full rounded-md bg-muted" />
-            <div className="h-5 w-[42%] rounded-md bg-muted" />
-          </div>
-        </div>
-
-        <div className="mt-auto">
-          <div className="rounded-2xl border border-border bg-muted p-4">
-            <div className="animate-pulse space-y-3">
-              <div className="h-6 w-full rounded-lg bg-muted" />
-              <div className="flex items-center justify-between gap-3">
-                <div className="flex items-center gap-2">
-                  <div className="size-8 rounded-full bg-muted" />
-                  <div className="size-8 rounded-full bg-muted" />
-                </div>
-                <div className="flex items-center gap-2">
-                  <div className="size-8 rounded-full bg-muted" />
-                  <div className="size-8 rounded-full bg-muted" />
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function AssistantTurnOutputs({
-  outputs,
-  onOpenOutput,
-  onOpenAllArtifacts,
-}: {
-  outputs: WorkspaceOutputRecordPayload[];
-  onOpenOutput?: (output: WorkspaceOutputRecordPayload) => void;
-  onOpenAllArtifacts: (outputs: WorkspaceOutputRecordPayload[]) => void;
-}) {
-  const displayOutputs =
-    outputs.length > 1 ? dedupeOutputsForDisplay(outputs) : outputs;
-  return (
-    <div className="mt-3 flex flex-col gap-2">
-      {displayOutputs.map((output) => (
-        <button
-          key={output.id}
-          type="button"
-          onClick={() => onOpenOutput?.(output)}
-          className="group flex max-w-[380px] items-center gap-3 rounded-xl border border-border bg-card px-3 py-2.5 text-left transition-colors hover:border-border hover:bg-accent/50 disabled:cursor-default disabled:hover:border-border disabled:hover:bg-card"
-          disabled={!onOpenOutput}
-        >
-          <OutputArtifactIcon output={output} />
-          <div className="min-w-0 flex-1">
-            <div className="truncate text-sm font-medium text-foreground">
-              {output.title || "Untitled artifact"}
-            </div>
-            <div className="truncate text-xs text-muted-foreground">
-              {outputSecondaryLabel(output)}
-            </div>
-          </div>
-          <ArrowUpRight className="size-3.5 shrink-0 text-muted-foreground/0 transition-colors group-hover:text-muted-foreground" />
-        </button>
-      ))}
-
-      {displayOutputs.length > 1 ? (
-        <button
-          type="button"
-          onClick={() => onOpenAllArtifacts(displayOutputs)}
-          className="flex max-w-[380px] items-center gap-3 rounded-xl border border-dashed border-border px-3 py-2 text-left text-muted-foreground transition-colors hover:border-border hover:bg-accent/40 hover:text-foreground"
-        >
-          <div className="grid size-7 shrink-0 place-items-center rounded-md bg-muted text-muted-foreground">
-            <Folder className="size-3.5" />
-          </div>
-          <span className="text-xs">
-            View artifacts in this reply ({displayOutputs.length})
-          </span>
-        </button>
-      ) : null}
-    </div>
-  );
-}
-
-function memoryProposalStateLabel(state: MemoryUpdateProposalState) {
-  switch (state) {
-    case "accepted":
-      return "Saved";
-    case "dismissed":
-      return "Dismissed";
-    default:
-      return "Review";
-  }
-}
-
-function AssistantTurnMemoryProposals({
-  proposals,
-  proposalAction,
-  editingProposalId,
-  drafts,
-  onEditProposal,
-  onDraftChange,
-  onAcceptProposal,
-  onDismissProposal,
-}: {
-  proposals: MemoryUpdateProposalRecordPayload[];
-  proposalAction: { proposalId: string; action: "accept" | "dismiss" } | null;
-  editingProposalId: string | null;
-  drafts: Record<string, string>;
-  onEditProposal: (proposalId: string) => void;
-  onDraftChange: (proposalId: string, value: string) => void;
-  onAcceptProposal: (proposal: MemoryUpdateProposalRecordPayload) => void;
-  onDismissProposal: (proposal: MemoryUpdateProposalRecordPayload) => void;
-}) {
-  return (
-    <div className="mt-4 grid gap-3">
-      {proposals.map((proposal) => {
-        const isPending = proposal.state === "pending";
-        const isEditing = editingProposalId === proposal.proposal_id;
-        const isActing = proposalAction?.proposalId === proposal.proposal_id;
-        const draftValue = drafts[proposal.proposal_id] ?? proposal.summary;
-
-        return (
-          <article
-            key={proposal.proposal_id}
-            className="bg-card rounded-2xl border border-border px-4 py-4 shadow-sm"
-          >
-            <div className="flex items-start justify-between gap-3">
-              <div className="min-w-0 flex-1">
-                <div className="flex items-center gap-2 text-sm font-medium text-foreground">
-                  <Lightbulb className="size-4 shrink-0 text-primary" />
-                  <span>{proposal.title}</span>
-                </div>
-                {isEditing ? (
-                  <textarea
-                    value={draftValue}
-                    onChange={(event) =>
-                      onDraftChange(proposal.proposal_id, event.target.value)
-                    }
-                    className="bg-muted mt-3 min-h-[86px] w-full rounded-xl border border-border px-3 py-2 text-sm leading-6 text-foreground outline-none transition focus:border-primary"
-                  />
-                ) : (
-                  <div className="mt-3 text-sm leading-6 text-muted-foreground">
-                    {proposal.summary}
-                  </div>
-                )}
-                {proposal.evidence ? (
-                  <div className="mt-3 text-xs leading-5 text-muted-foreground">
-                    {proposal.evidence}
-                  </div>
-                ) : null}
-              </div>
-
-              <div className="flex shrink-0 items-start gap-2">
-                <Badge variant="outline" className="uppercase">
-                  {memoryProposalStateLabel(proposal.state)}
-                </Badge>
-                {isPending ? (
-                  <Button
-                    type="button"
-                    variant="outline"
-                    size="icon"
-                    onClick={() => onEditProposal(proposal.proposal_id)}
-                    className="rounded-xl"
-                    aria-label="Edit memory proposal"
-                  >
-                    <PencilLine className="size-3.5" />
-                  </Button>
-                ) : null}
-              </div>
-            </div>
-
-            {isPending ? (
-              <div className="mt-4 flex flex-wrap gap-2">
-                <Button
-                  type="button"
-                  variant="outline"
-                  size="lg"
-                  onClick={() => onDismissProposal(proposal)}
-                  disabled={isActing}
-                  className="rounded-2xl"
-                >
-                  {isActing && proposalAction?.action === "dismiss" ? (
-                    <Loader2 className="size-3 animate-spin" />
-                  ) : (
-                    <X className="size-3" />
-                  )}
-                  <span>Dismiss</span>
-                </Button>
-                <Button
-                  type="button"
-                  variant="outline"
-                  size="lg"
-                  onClick={() => onAcceptProposal(proposal)}
-                  disabled={isActing}
-                  className="rounded-2xl border-primary bg-primary/10 text-primary hover:bg-primary/14"
-                >
-                  {isActing && proposalAction?.action === "accept" ? (
-                    <Loader2 className="size-3 animate-spin" />
-                  ) : (
-                    <Check className="size-3" />
-                  )}
-                  <span>Accept</span>
-                </Button>
-              </div>
-            ) : null}
-          </article>
-        );
-      })}
-    </div>
-  );
-}
-
-export function ConversationTurns<Message extends ChatMessage>({
-  messages,
-  assistantLabel,
-  assistantMode,
-  showExecutionInternals,
-  assistantFitToContent = false,
-  /** Drives the agent avatar's seed so each workspace has its own
-   *  persistent face. */
-  workspaceId,
-  onPreviewAttachment,
-  onOpenOutput,
-  onOpenAllArtifacts,
-  collapsedTraceByStepId,
-  onToggleTraceStep,
-  onLinkClick,
-  onLocalLinkClick,
-  memoryProposalAction,
-  editingMemoryProposalId,
-  memoryProposalDrafts,
-  onEditMemoryProposal,
-  onMemoryProposalDraftChange,
-  onAcceptMemoryProposal,
-  onDismissMemoryProposal,
-  assistantFooterAccessoryMessageId = null,
-  assistantFooterAccessory = null,
-  getMessageWrapperClassName,
-  liveAssistantTurn = null,
-}: {
-  messages: Message[];
-  assistantLabel: string;
-  assistantMode: string;
-  showExecutionInternals: boolean;
-  assistantFitToContent?: boolean;
-  workspaceId?: string | null;
-  onPreviewAttachment?: (attachment: AttachmentListItem) => void;
-  onOpenOutput?: (output: WorkspaceOutputRecordPayload) => void;
-  onOpenAllArtifacts: (outputs: WorkspaceOutputRecordPayload[]) => void;
-  collapsedTraceByStepId: Record<string, boolean>;
-  onToggleTraceStep: (stepId: string) => void;
-  onLinkClick?: (url: string) => void;
-  onLocalLinkClick?: (href: string) => void;
-  memoryProposalAction: {
-    proposalId: string;
-    action: "accept" | "dismiss";
-  } | null;
-  editingMemoryProposalId: string | null;
-  memoryProposalDrafts: Record<string, string>;
-  onEditMemoryProposal: (message: Message, proposalId: string) => void;
-  onMemoryProposalDraftChange: (proposalId: string, value: string) => void;
-  onAcceptMemoryProposal: (proposal: MemoryUpdateProposalRecordPayload) => void;
-  onDismissMemoryProposal: (
-    proposal: MemoryUpdateProposalRecordPayload,
-  ) => void;
-  assistantFooterAccessoryMessageId?: string | null;
-  assistantFooterAccessory?: ReactNode;
-  getMessageWrapperClassName?: (message: Message) => string | undefined;
-  liveAssistantTurn?: {
-    text: string;
-    tone?: ChatMessage["tone"];
-    segments: ChatAssistantSegment[];
-    executionItems: ChatExecutionTimelineItem[];
-    status?: string;
-    statusAccessory?: ReactNode;
-    footerAccessory?: ReactNode;
-  } | null;
-}) {
-  return (
-    <>
-      {messages.map((message, index) => {
-        const wrapperClassName = getMessageWrapperClassName?.(message)?.trim();
-        const next = messages[index + 1];
-        const isLastInAssistantGroup =
-          message.role === "assistant" &&
-          (!next || next.role === "user") &&
-          !liveAssistantTurn;
-        const turn =
-          message.role === "user" ? (
-            <UserTurn
-              text={message.text}
-              createdAt={message.createdAt}
-              attachments={message.attachments ?? []}
-              onPreviewAttachment={onPreviewAttachment}
-              onLinkClick={onLinkClick}
-              onLocalLinkClick={onLocalLinkClick}
-            />
-          ) : (
-            <AssistantTurn
-              label={assistantLabel}
-              mode={assistantMode}
-              showExecutionInternals={showExecutionInternals}
-              fitToContent={assistantFitToContent}
-              text={message.text}
-              tone={message.tone ?? "default"}
-              segments={message.segments ?? []}
-              executionItems={message.executionItems ?? []}
-              memoryProposals={message.memoryProposals ?? []}
-              outputs={message.outputs ?? []}
-              memoryProposalAction={memoryProposalAction}
-              editingMemoryProposalId={editingMemoryProposalId}
-              memoryProposalDrafts={memoryProposalDrafts}
-              onEditMemoryProposal={(proposalId) =>
-                onEditMemoryProposal(message, proposalId)
-              }
-              onMemoryProposalDraftChange={onMemoryProposalDraftChange}
-              onAcceptMemoryProposal={onAcceptMemoryProposal}
-              onDismissMemoryProposal={onDismissMemoryProposal}
-              onOpenOutput={onOpenOutput}
-              onOpenAllArtifacts={onOpenAllArtifacts}
-              collapsedTraceByStepId={collapsedTraceByStepId}
-              onToggleTraceStep={onToggleTraceStep}
-              onLinkClick={onLinkClick}
-              onLocalLinkClick={onLocalLinkClick}
-              showAvatar={isLastInAssistantGroup}
-              workspaceId={workspaceId ?? null}
-              createdAt={message.createdAt}
-              footerAccessory={
-                message.id === assistantFooterAccessoryMessageId
-                  ? assistantFooterAccessory
-                  : null
-              }
-            />
-          );
-
-        if (wrapperClassName) {
-          return (
-            <div key={message.id} className={wrapperClassName}>
-              {turn}
-            </div>
-          );
-        }
-        return <Fragment key={message.id}>{turn}</Fragment>;
-      })}
-
-      {liveAssistantTurn ? (
-        <AssistantTurn
-          label={assistantLabel}
-          mode={assistantMode}
-          showExecutionInternals={showExecutionInternals}
-          fitToContent={assistantFitToContent}
-          text={liveAssistantTurn.text}
-          tone={liveAssistantTurn.tone ?? "default"}
-          segments={liveAssistantTurn.segments}
-          executionItems={liveAssistantTurn.executionItems}
-          memoryProposals={[]}
-          outputs={[]}
-          memoryProposalAction={memoryProposalAction}
-          editingMemoryProposalId={editingMemoryProposalId}
-          memoryProposalDrafts={memoryProposalDrafts}
-          onEditMemoryProposal={() => undefined}
-          onMemoryProposalDraftChange={onMemoryProposalDraftChange}
-          onAcceptMemoryProposal={onAcceptMemoryProposal}
-          onDismissMemoryProposal={onDismissMemoryProposal}
-          onOpenOutput={onOpenOutput}
-          onOpenAllArtifacts={onOpenAllArtifacts}
-          collapsedTraceByStepId={collapsedTraceByStepId}
-          onToggleTraceStep={onToggleTraceStep}
-          onLinkClick={onLinkClick}
-          onLocalLinkClick={onLocalLinkClick}
-          showAvatar
-          workspaceId={workspaceId ?? null}
-          live
-          statusAccessory={liveAssistantTurn.statusAccessory ?? null}
-          status={liveAssistantTurn.status ?? ""}
-          footerAccessory={liveAssistantTurn.footerAccessory ?? null}
-        />
-      ) : null}
-    </>
-  );
-}
-
-export function ArtifactBrowserModal({
-  open,
-  filter,
-  outputs,
-  scope,
-  onClose,
-  onFilterChange,
-  onOpenOutput,
-  layout = "page",
-}: {
-  open: boolean;
-  filter: ArtifactBrowserFilter;
-  outputs: WorkspaceOutputRecordPayload[];
-  scope: "session" | "reply";
-  onClose: () => void;
-  onFilterChange: (nextFilter: ArtifactBrowserFilter) => void;
-  onOpenOutput?: (output: WorkspaceOutputRecordPayload) => void;
-  layout?: "page" | "card";
-}) {
-  if (!open) {
-    return null;
-  }
-
-  const filterLabels: Array<{ id: ArtifactBrowserFilter; label: string }> = [
-    { id: "all", label: "All" },
-    { id: "documents", label: "Documents" },
-    { id: "images", label: "Images" },
-    { id: "code", label: "Code files" },
-    { id: "links", label: "Links" },
-    { id: "apps", label: "Apps" },
-  ];
-  const allDisplayOutputs =
-    outputs.length > 1 ? dedupeOutputsForDisplay(outputs) : outputs;
-  const filteredOutputs = sortOutputsLatestFirst(
-    filter === "all"
-      ? allDisplayOutputs
-      : allDisplayOutputs.filter(
-          (output) => outputBrowserFilterForOutput(output) === filter,
-        ),
-  );
-  const overlayClassName =
-    layout === "card"
-      ? "absolute inset-0 z-30 flex items-stretch justify-stretch bg-background/88 p-2 backdrop-blur-[2px]"
-      : "absolute inset-0 z-30 flex items-center justify-center bg-black/40 px-6 py-8 backdrop-blur-[2px]";
-  const panelClassName =
-    layout === "card"
-      ? "flex h-full w-full min-h-0 flex-col overflow-hidden rounded-2xl border border-border bg-background shadow-xl"
-      : "flex max-h-full w-full max-w-lg flex-col overflow-hidden rounded-xl border border-border bg-background shadow-xl";
-
-  return (
-    <div
-      className={overlayClassName}
-      data-control-center-swipe-ignore={layout === "card" ? "true" : undefined}
-    >
-      <div className={panelClassName}>
-        <div className="flex shrink-0 items-center justify-between gap-3 border-b border-border px-4 py-3">
-          <div>
-            <div className="text-sm font-semibold text-foreground">
-              Artifacts
-            </div>
-            <div className="text-xs text-muted-foreground">
-              {allDisplayOutputs.length} item
-              {allDisplayOutputs.length === 1 ? "" : "s"}{" "}
-              {scope === "reply"
-                ? "attached to this reply"
-                : "in this session"}
-            </div>
-          </div>
-          <Button
-            type="button"
-            variant="ghost"
-            size="icon-sm"
-            onClick={onClose}
-            aria-label="Close"
-          >
-            <X className="size-3.5" />
-          </Button>
-        </div>
-
-        <div className="flex shrink-0 flex-wrap gap-1 border-b border-border px-3 py-2">
-          {filterLabels.map((item) => {
-            const active = filter === item.id;
-            return (
-              <Button
-                key={item.id}
-                variant={active ? "secondary" : "ghost"}
-                size="xs"
-                onClick={() => onFilterChange(item.id)}
-              >
-                {item.label}
-              </Button>
-            );
-          })}
-        </div>
-
-        <div className="min-h-0 flex-1 overflow-y-auto px-3 py-3">
-          {filteredOutputs.length === 0 ? (
-            <div className="py-10 text-center text-xs text-muted-foreground">
-              No artifacts match this filter.
-            </div>
-          ) : (
-            <div className="grid gap-2">
-              {filteredOutputs.map((output) => {
-                const kindLabel = outputKindLabel(output);
-                const changeLabel = outputChangeLabel(output);
-                return (
-                  <button
-                    key={output.id}
-                    type="button"
-                    onClick={() => {
-                      onClose();
-                      onOpenOutput?.(output);
-                    }}
-                    disabled={!onOpenOutput}
-                    className="group flex w-full min-w-0 items-start gap-2.5 rounded-lg border border-border bg-card px-3 py-2.5 text-left ring-border transition-colors hover:bg-accent/50 disabled:cursor-default disabled:hover:bg-card"
-                  >
-                    <OutputArtifactIcon output={output} />
-                    <div className="min-w-0 flex-1">
-                      <div className="text-[10px] font-medium uppercase tracking-wide text-muted-foreground">
-                        {kindLabel}
-                      </div>
-                      <div className="truncate text-sm font-medium text-foreground">
-                        {output.title || "Untitled artifact"}
-                      </div>
-                    </div>
-                    {changeLabel ? (
-                      <Badge
-                        variant="outline"
-                        className="shrink-0 text-[10px] uppercase"
-                      >
-                        {changeLabel}
-                      </Badge>
-                    ) : null}
-                  </button>
-                );
-              })}
-            </div>
-          )}
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function traceStatusLabel(status: ChatTraceStepStatus) {
-  if (status === "completed") {
-    return "Completed";
-  }
-  if (status === "error") {
-    return "Error";
-  }
-  if (status === "waiting") {
-    return "Waiting";
-  }
-  return "In progress";
-}
-
-function isTraceStepCollapsed(
-  step: ChatTraceStep,
-  collapsedTraceByStepId: Record<string, boolean>,
-) {
-  return collapsedTraceByStepId[step.id] ?? true;
-}
-
-function IntegrationErrorBanner({ details }: { details: string[] }) {
-  const errorText = details.join(" ");
-  const integrationError = isIntegrationError(errorText);
-  if (!integrationError) return null;
-  return (
-    <div className="mt-1.5 flex items-center gap-2 rounded-lg border border-warning/20 bg-warning/8 px-2.5 py-1.5 text-xs text-warning">
-      <Cable className="size-3 shrink-0" />
-      <span>{integrationError.action}</span>
-    </div>
-  );
-}
-
-function LiveStatusEllipsis() {
-  return (
-    <span
-      aria-hidden="true"
-      className="inline-flex shrink-0 items-center text-muted-foreground"
-    >
-      <DotmSquare3 dotSize={1} size={10} />
-    </span>
-  );
-}
-
-function StreamingCursor() {
-  return (
-    <>
-      <style>{`
-        @keyframes streaming-cursor-blink {
-          0%, 50% { opacity: 1; }
-          50.01%, 100% { opacity: 0; }
-        }
-      `}</style>
-      <span
-        aria-hidden="true"
-        className="ml-0.5 inline-block h-[1em] w-[2px] -mb-[2px] translate-y-[3px] rounded-[1px] bg-foreground/65"
-        style={{ animation: "streaming-cursor-blink 1100ms steps(1) infinite" }}
-      />
-    </>
-  );
-}
-
-function LiveStatusLine({
-  label,
-  className = "",
-}: {
-  label: string;
-  className?: string;
-}) {
-  const normalizedLabel = label.replace(/\.+$/, "").trim();
-  if (!normalizedLabel) {
-    return null;
-  }
-
-  return (
-    <div
-      aria-live="polite"
-      key={normalizedLabel}
-      className={`flex w-fit items-center gap-1.5 text-xs leading-none text-muted-foreground animate-in fade-in-0 slide-in-from-bottom-0.5 duration-200 ease-out ${className}`.trim()}
-    >
-      <LiveStatusEllipsis />
-      <span>{normalizedLabel}</span>
-    </div>
-  );
-}
-
-function summarizeThinking(text: string) {
-  const firstContentLine =
-    text
-      .split("\n")
-      .map((line) => line.replace(/[*_`#>-]/g, "").trim())
-      .find(Boolean) || "Reasoning available";
-
-  return firstContentLine.length > 88
-    ? `${firstContentLine.slice(0, 85).trimEnd()}...`
-    : firstContentLine;
-}
-
-function TraceTimelineStepEntry({
-  step,
-  collapsedByStepId,
-  onToggleStep,
-}: {
-  step: ChatTraceStep;
-  collapsedByStepId: Record<string, boolean>;
-  onToggleStep: (stepId: string) => void;
-}) {
-  const expanded = !(collapsedByStepId[step.id] ?? true);
-
-  return (
-    <div>
-      <button
-        type="button"
-        onClick={() => step.details.length > 0 && onToggleStep(step.id)}
-        className={`flex w-full items-start gap-2 rounded-md px-2.5 -ml-2.5 py-1 text-left text-xs transition-colors ${step.details.length > 0 ? "hover:bg-muted cursor-pointer" : "cursor-default"}`}
-      >
-        <span className="mt-0.5 shrink-0">
-          {step.status === "completed" ? (
-            <Check className="size-3 text-success" />
-          ) : step.status === "error" ? (
-            <AlertTriangle className="size-3 text-destructive" />
-          ) : step.status === "running" ? (
-            <Loader2 className="size-3 animate-spin text-muted-foreground" />
-          ) : (
-            <Clock3 className="size-3 text-muted-foreground" />
-          )}
-        </span>
-        <span className="min-w-0 flex-1">
-          <span className="font-medium text-foreground">{step.title}</span>
-          {step.details.length > 0 ? (
-            <span className="ml-1.5 text-muted-foreground">
-              {step.details[0]}
-            </span>
-          ) : null}
-        </span>
-        {step.details.length > 1 ? (
-          <ChevronDown
-            className={`size-3 mt-0.5 shrink-0 text-muted-foreground transition-transform ${expanded ? "rotate-180" : ""}`}
-          />
-        ) : null}
-      </button>
-      {expanded && step.details.length > 1 ? (
-        <div className="ml-6 mt-0.5 mb-1 rounded-md border border-border bg-muted px-3 py-2 text-xs leading-5 text-muted-foreground whitespace-pre-wrap">
-          {step.details.slice(1).join("\n")}
-        </div>
-      ) : null}
-      {step.status === "error" ? (
-        <IntegrationErrorBanner details={step.details} />
-      ) : null}
-    </div>
-  );
-}
-
-function ExecutionTimelineThinkingEntry({
-  text,
-  onLinkClick,
-  onLocalLinkClick,
-}: {
-  text: string;
-  onLinkClick?: (url: string) => void;
-  onLocalLinkClick?: (href: string) => void;
-}) {
-  return (
-    <div className="py-1">
-      <div className="-ml-2.5 w-[calc(100%+0.625rem)] rounded-xl border border-border bg-muted px-3.5 py-3">
-        <SimpleMarkdown
-          className="chat-markdown chat-thinking-markdown max-w-full text-foreground"
-          onLinkClick={onLinkClick}
-          onLocalLinkClick={onLocalLinkClick}
-        >
-          {text}
-        </SimpleMarkdown>
-      </div>
-    </div>
-  );
-}
-
-function TraceStepGroup({
-  items,
-  collapsedByStepId,
-  onToggleStep,
-  live = false,
-  liveOutputStarted = false,
-  onLinkClick,
-  onLocalLinkClick,
-  forceExpandToken = 0,
-}: {
-  items: ChatExecutionTimelineItem[];
-  collapsedByStepId: Record<string, boolean>;
-  onToggleStep: (stepId: string) => void;
-  live?: boolean;
-  liveOutputStarted?: boolean;
-  onLinkClick?: (url: string) => void;
-  onLocalLinkClick?: (href: string) => void;
-  forceExpandToken?: number;
-}) {
-  const steps = traceStepsFromExecutionItems(items);
-  const [groupExpanded, setGroupExpanded] = useState(
-    live && !liveOutputStarted,
-  );
-  const previousLiveRef = useRef(live);
-  const previousLiveOutputStartedRef = useRef(liveOutputStarted);
-  const previousForceExpandTokenRef = useRef(forceExpandToken);
-
-  useEffect(() => {
-    if (forceExpandToken !== previousForceExpandTokenRef.current) {
-      previousForceExpandTokenRef.current = forceExpandToken;
-      if (forceExpandToken > 0) {
-        setGroupExpanded(true);
-      }
-    }
-  }, [forceExpandToken]);
-
-  useEffect(() => {
-    if (live && !previousLiveRef.current) {
-      setGroupExpanded(!liveOutputStarted);
-    }
-    if (live && liveOutputStarted && !previousLiveOutputStartedRef.current) {
-      setGroupExpanded(false);
-    }
-    previousLiveRef.current = live;
-    previousLiveOutputStartedRef.current = liveOutputStarted;
-  }, [live, liveOutputStarted]);
-  const runningCount = steps.filter((s) => s.status === "running").length;
-  const terminalErrorCount = steps.filter(
-    (step) => step.kind === "phase" && step.status === "error",
-  ).length;
-  const groupHasTerminalError = terminalErrorCount > 0;
-  const stepCount = steps.length;
-  const stepLabel = `${stepCount} step${stepCount === 1 ? "" : "s"}`;
-  const activeStep =
-    [...steps]
-      .reverse()
-      .find((step) => step.status === "running" || step.status === "waiting") ??
-    null;
-  const groupIsLive = live && activeStep !== null && !groupHasTerminalError;
-  const latestStep = steps.length > 0 ? steps[steps.length - 1] : null;
-  const latestThinkingItem =
-    [...items]
-      .reverse()
-      .find(
-        (
-          item,
-        ): item is Extract<ChatExecutionTimelineItem, { kind: "thinking" }> =>
-          item.kind === "thinking",
-      ) ?? null;
-  const summaryStep = activeStep ?? (groupIsLive ? latestStep : null);
-  const summarySuffix = groupHasTerminalError
-    ? ` (${terminalErrorCount} failed)`
-    : "";
-  const showLiveSummarySpinner =
-    (groupIsLive || runningCount > 0) && !groupExpanded;
-  const summaryLabel = summaryStep
-    ? summaryStep === activeStep || summaryStep.status === "waiting"
-      ? `${traceStatusLabel(summaryStep.status)}: ${summaryStep.title}`
-      : groupIsLive
-        ? summaryStep.title
-        : `${traceStatusLabel(summaryStep.status)}: ${summaryStep.title}`
-    : groupIsLive
-      ? latestThinkingItem
-        ? summarizeThinking(latestThinkingItem.text)
-        : stepCount > 0
-          ? `Working through ${stepLabel}...`
-          : "Thinking..."
-      : runningCount > 0
-        ? `Running ${stepLabel}...`
-        : latestThinkingItem
-          ? summarizeThinking(latestThinkingItem.text)
-          : stepCount > 0 && latestStep
-            ? latestStep.title
-            : "Execution trace";
-
-  return (
-    <div className="mt-3 first:mt-0">
-      <button
-        type="button"
-        onClick={() => setGroupExpanded((v) => !v)}
-        className="flex w-full items-center gap-2 rounded-lg px-2.5 py-1.5 -ml-2.5 text-left text-xs text-muted-foreground transition-colors hover:bg-muted"
-      >
-        {groupHasTerminalError ? (
-          <AlertTriangle className="size-3.5 shrink-0 text-destructive" />
-        ) : showLiveSummarySpinner ? (
-          <Loader2 className="size-3.5 shrink-0 animate-spin text-muted-foreground" />
-        ) : groupIsLive || runningCount > 0 ? (
-          <Clock3 className="size-3.5 shrink-0 text-muted-foreground" />
-        ) : (
-          <Check className="size-3.5 shrink-0 text-success" />
-        )}
-        <span className="min-w-0 flex-1 truncate leading-5">
-          {summaryLabel}
-          {summarySuffix}
-        </span>
-        {stepCount > 0 && !groupIsLive && !groupHasTerminalError ? (
-          <span
-            aria-hidden
-            className="shrink-0 rounded-full bg-muted px-1.5 py-px text-[10px] tabular-nums text-muted-foreground"
-          >
-            {stepCount}
-          </span>
-        ) : null}
-        <ChevronDown
-          className={`size-3 shrink-0 transition-transform ${groupExpanded ? "rotate-180" : ""}`}
-        />
-      </button>
-
-      {groupExpanded ? (
-        <div className="mt-1 ml-1 space-y-0.5">
-          {items.map((item) =>
-            item.kind === "thinking" ? (
-              <ExecutionTimelineThinkingEntry
-                key={item.id}
-                text={item.text}
-                onLinkClick={onLinkClick}
-                onLocalLinkClick={onLocalLinkClick}
-              />
-            ) : (
-              <TraceTimelineStepEntry
-                key={item.id}
-                step={item.step}
-                collapsedByStepId={collapsedByStepId}
-                onToggleStep={onToggleStep}
-              />
-            ),
-          )}
-        </div>
-      ) : null}
-    </div>
-  );
-}
-
-function AttachmentImageThumb({ file }: { file: File }) {
-  const [objectUrl, setObjectUrl] = useState<string | null>(null);
-  const [loadFailed, setLoadFailed] = useState(false);
-
-  useEffect(() => {
-    const url = URL.createObjectURL(file);
-    setObjectUrl(url);
-    return () => URL.revokeObjectURL(url);
-  }, [file]);
-
-  if (loadFailed || !objectUrl) {
-    return <ImageIcon className="size-4 shrink-0 text-primary" />;
-  }
-  return (
-    <img
-      alt=""
-      className="size-7 shrink-0 rounded-md object-cover"
-      onError={() => setLoadFailed(true)}
-      src={objectUrl}
-    />
-  );
-}
-
-function AttachmentList({
-  attachments,
-  onRemove,
-  onPreview,
-  className = "",
-}: {
-  attachments: AttachmentListItem[];
-  onRemove?: (attachmentId: string) => void;
-  onPreview?: (attachment: AttachmentListItem) => void;
-  className?: string;
-}) {
-  return (
-    <div className={`flex flex-wrap gap-2 ${className}`.trim()}>
-      {attachments.map((attachment) => {
-        const isImagePreviewable =
-          attachment.kind === "image" &&
-          Boolean(onPreview) &&
-          Boolean(
-            attachment.file ||
-            (typeof attachment.workspace_path === "string" &&
-              attachment.workspace_path.trim()),
-          );
-
-        const icon =
-          attachment.kind === "image" && attachment.file ? (
-            <AttachmentImageThumb file={attachment.file} />
-          ) : attachment.kind === "image" ? (
-            <ImageIcon className="size-4 shrink-0 text-primary" />
-          ) : attachment.kind === "folder" ? (
-            <Folder className="size-3.5 shrink-0 text-primary" />
-          ) : (
-            <FileText className="size-3.5 shrink-0 text-primary" />
-          );
-
-        const labelClassName = "truncate";
-        const isImageThumb = attachment.kind === "image" && Boolean(attachment.file);
-
-        const content = (
-          <>
-            {icon}
-            <span className={labelClassName}>
-              {attachmentButtonLabel(attachment)}
-            </span>
-          </>
-        );
-
-        return (
-          <div
-            className="group/attachment bg-muted relative inline-flex max-w-full items-center gap-2 rounded-lg border border-border pr-2 text-xs text-foreground"
-            key={attachment.id}
-            style={{
-              paddingLeft: isImageThumb ? "4px" : "10px",
-              paddingTop: isImageThumb ? "4px" : "5px",
-              paddingBottom: isImageThumb ? "4px" : "5px",
-            }}
-          >
-            {isImagePreviewable ? (
-              <button
-                aria-label={`Preview ${attachment.name}`}
-                className="flex min-w-0 items-center gap-2 rounded-md text-left transition hover:text-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-primary/40"
-                onClick={() => onPreview?.(attachment)}
-                title={`Preview ${attachment.name}`}
-                type="button"
-              >
-                {content}
-              </button>
-            ) : (
-              content
-            )}
-            {onRemove ? (
-              <button
-                aria-label={`Remove ${attachment.name}`}
-                className="grid size-4 shrink-0 place-items-center rounded-full text-muted-foreground opacity-0 transition group-hover/attachment:opacity-100 hover:text-foreground"
-                onClick={() => onRemove(attachment.id)}
-                type="button"
-              >
-                <X className="size-3" />
-              </button>
-            ) : null}
-          </div>
-        );
-      })}
-    </div>
-  );
-}
-
-function ImageAttachmentPreviewModal({
-  open,
-  preview,
-  onClose,
-}: {
-  open: boolean;
-  preview: ImageAttachmentPreviewState | null;
-  onClose: () => void;
-}) {
-  useEffect(() => {
-    if (!open) {
-      return;
-    }
-
-    const handleKeyDown = (event: globalThis.KeyboardEvent) => {
-      if (event.key !== "Escape") {
-        return;
-      }
-      event.preventDefault();
-      onClose();
-    };
-
-    window.addEventListener("keydown", handleKeyDown);
-    return () => {
-      window.removeEventListener("keydown", handleKeyDown);
-    };
-  }, [open, onClose]);
-
-  if (!open || !preview) {
-    return null;
-  }
-
-  const sizeLabel = formatAttachmentSize(preview.attachment.size_bytes);
-  const showImage = !preview.isLoading && !preview.errorMessage;
-  const modalContent = (
-    <div
-      className="fixed inset-0 z-[90] flex items-center justify-center px-6 py-8"
-      onClick={onClose}
-    >
-      {preview.browserSnapshot ? (
-        <img
-          aria-hidden="true"
-          src={preview.browserSnapshot.dataUrl}
-          alt=""
-          className="pointer-events-none absolute object-fill"
-          style={{
-            left: `${preview.browserSnapshot.bounds.x}px`,
-            top: `${preview.browserSnapshot.bounds.y}px`,
-            width: `${preview.browserSnapshot.bounds.width}px`,
-            height: `${preview.browserSnapshot.bounds.height}px`,
-          }}
-        />
-      ) : null}
-      <div
-        aria-hidden="true"
-        className="absolute inset-0 bg-black/70 backdrop-blur-[2px]"
-      />
-      <div
-        role="dialog"
-        aria-modal="true"
-        aria-label={`Preview ${preview.attachment.name}`}
-        className="relative z-10 flex max-h-[calc(100vh-64px)] flex-col overflow-hidden rounded-2xl border border-white/10 bg-background shadow-2xl"
-        style={{ maxWidth: "92vw" }}
-        onClick={(event) => event.stopPropagation()}
-      >
-        <div className="flex shrink-0 items-center justify-between gap-3 border-b border-border px-4 py-3">
-          <div className="min-w-0">
-            <div className="truncate text-sm font-semibold text-foreground">
-              {preview.attachment.name}
-            </div>
-            <div className="truncate text-xs text-muted-foreground">
-              {sizeLabel || "Image attachment"}
-            </div>
-          </div>
-          <Button
-            type="button"
-            variant="ghost"
-            size="icon-sm"
-            onClick={onClose}
-            aria-label="Close image preview"
-          >
-            <X className="size-3.5" />
-          </Button>
-        </div>
-
-        <div
-          className={`overflow-auto px-4 py-4 ${
-            showImage
-              ? "bg-transparent"
-              : "min-h-[240px] min-w-[320px] bg-muted/20"
-          }`}
-        >
-          {preview.isLoading ? (
-            <div className="flex h-full min-h-[208px] items-center justify-center gap-2 text-sm text-foreground/80">
-              <Loader2 className="size-4 animate-spin" />
-              <span>Loading preview...</span>
-            </div>
-          ) : preview.errorMessage ? (
-            <div className="flex h-full min-h-[208px] flex-col items-center justify-center gap-3 px-6 text-center">
-              <AlertTriangle className="size-5 text-warning" />
-              <p className="max-w-md text-sm text-foreground/80">
-                {preview.errorMessage}
-              </p>
-            </div>
-          ) : (
-            <div className="flex items-center justify-center">
-              <img
-                src={preview.dataUrl}
-                alt={preview.attachment.name}
-                className="block h-auto w-auto rounded-lg ring-1 ring-black/8"
-                style={{
-                  maxWidth: "calc(92vw - 32px)",
-                  maxHeight: "calc(88vh - 128px)",
-                }}
-              />
-            </div>
-          )}
-        </div>
-      </div>
-    </div>
-  );
-
-  return createPortal(modalContent, document.body);
-}
-
-function ModelCombobox({
-  selectedModel,
-  selectedModelLabel,
-  runtimeDefaultModelLabel,
-  runtimeDefaultModelAvailable,
-  modelOptions,
-  modelOptionGroups,
-  disabled,
-  compact = false,
-  onModelChange,
-}: {
-  selectedModel: string;
-  selectedModelLabel: string;
-  runtimeDefaultModelLabel: string;
-  runtimeDefaultModelAvailable: boolean;
-  modelOptions: ChatModelOption[];
-  modelOptionGroups: ChatModelOptionGroup[];
-  disabled: boolean;
-  compact?: boolean;
-  onModelChange: (value: string) => void;
-}) {
-  const [open, setOpen] = useState(false);
-  const [query, setQuery] = useState("");
-
-  const autoOption = useMemo(
-    () =>
-      runtimeDefaultModelAvailable
-        ? ({
-            value: CHAT_MODEL_USE_RUNTIME_DEFAULT,
-            label: `Auto (${runtimeDefaultModelLabel})`,
-          } satisfies ChatModelOption)
-        : null,
-    [runtimeDefaultModelAvailable, runtimeDefaultModelLabel],
-  );
-
-  const filteredAutoOption = useMemo(() => {
-    const q = query.trim().toLowerCase();
-    if (!autoOption) {
-      return null;
-    }
-    if (!q) {
-      return autoOption;
-    }
-    return autoOption.label.toLowerCase().includes(q) ||
-      autoOption.value.toLowerCase().includes(q)
-      ? autoOption
-      : null;
-  }, [autoOption, query]);
-
-  const filteredOptionGroups = useMemo(() => {
-    const q = query.trim().toLowerCase();
-    const sourceGroups =
-      modelOptionGroups.length > 0
-        ? modelOptionGroups
-        : [{ label: "", options: modelOptions }];
-    return sourceGroups
-      .map((group) => ({
-        ...group,
-        options: q
-          ? group.options.filter((option) => {
-              const haystack = [
-                option.label,
-                option.selectedLabel,
-                option.searchText,
-                option.value,
-                group.label,
-              ]
-                .filter(Boolean)
-                .join(" ")
-                .toLowerCase();
-              return haystack.includes(q);
-            })
-          : group.options,
-      }))
-      .filter((group) => group.options.length > 0);
-  }, [modelOptionGroups, modelOptions, query]);
-
-  const displayLabel =
-    selectedModel === CHAT_MODEL_USE_RUNTIME_DEFAULT
-      ? `Auto (${runtimeDefaultModelLabel})`
-      : selectedModelLabel || "Select model";
-  const compactLabel = compactComposerModelLabel(displayLabel);
-
-  const hasFilteredOptions =
-    Boolean(filteredAutoOption) ||
-    filteredOptionGroups.some((group) => group.options.length > 0);
-
-  const renderOption = (option: ChatModelOption) => {
-    const active = option.value === selectedModel;
-    const optionDisabled = Boolean(option.disabled);
-    // Auto/runtime-default doesn't represent a single model — keep its
-    // icon empty rather than guessing (the chosen runtime default still
-    // ends up rendering with its real brand mark in the trigger).
-    const isRuntimeDefault = option.value === CHAT_MODEL_USE_RUNTIME_DEFAULT;
-    return (
-      <button
-        key={option.value}
-        type="button"
-        disabled={optionDisabled}
-        aria-current={active ? "true" : undefined}
-        onClick={() => {
-          if (optionDisabled) {
-            return;
-          }
-          onModelChange(option.value);
-          setOpen(false);
-          setQuery("");
-        }}
-        className={`flex w-full items-center justify-between gap-2 rounded-md px-2.5 py-1.5 text-left text-xs transition-colors ${
-          active
-            ? "bg-accent text-foreground"
-            : optionDisabled
-              ? "cursor-not-allowed text-muted-foreground"
-              : "text-foreground hover:bg-accent"
-        }`}
-      >
-        <span className="flex min-w-0 items-center gap-2">
-          {isRuntimeDefault ? (
-            <span className="size-3.5 shrink-0" aria-hidden="true" />
-          ) : (
-            <ProviderBrandIcon
-              modelToken={option.value}
-              className="size-3.5 shrink-0"
-            />
-          )}
-          <span className="truncate">{option.label}</span>
-        </span>
-        {!active && option.statusLabel ? (
-          <span className="shrink-0 text-[10px] font-medium uppercase text-muted-foreground">
-            {option.statusLabel}
-          </span>
-        ) : null}
-      </button>
-    );
-  };
-
-  return (
-    <Popover
-      open={open}
-      onOpenChange={(nextOpen) => {
-        setOpen(nextOpen);
-        if (!nextOpen) setQuery("");
-      }}
-    >
-      <PopoverTrigger
-        disabled={disabled}
-        render={
-          <Button
-            variant="ghost"
-            size="sm"
-            className={`gap-1.5 rounded-md text-xs font-medium ${
-              compact ? "w-full justify-between px-2.5" : "px-2"
-            }`}
-          >
-            {compact ? (
-              <>
-                <span className="flex min-w-0 items-center gap-1.5">
-                  <ProviderBrandIcon
-                    modelToken={selectedModel}
-                    className="size-3.5 shrink-0"
-                  />
-                  <span className="truncate">{compactLabel}</span>
-                </span>
-                <ChevronDown className="size-3.5 shrink-0 text-muted-foreground" />
-              </>
-            ) : (
-              <>
-                <ProviderBrandIcon
-                  modelToken={selectedModel}
-                  className="size-3.5 shrink-0"
-                />
-                <span className="whitespace-nowrap">{displayLabel}</span>
-                <ChevronDown className="size-3.5 shrink-0 text-muted-foreground" />
-              </>
-            )}
-          </Button>
-        }
-      />
-      <PopoverContent
-        align="start"
-        side="top"
-        sideOffset={8}
-        className="p-0 gap-0"
-      >
-        <div className="border-b border-border p-1.5">
-          <div className="relative flex h-7 items-center rounded-md border border-border bg-background px-2.5">
-            <Search className="size-3.5 shrink-0 text-muted-foreground" />
-            <input
-              value={query}
-              onChange={(e) => setQuery(e.target.value)}
-              placeholder="Search models..."
-              className="embedded-input h-full w-full bg-transparent pl-2 text-xs text-foreground outline-none placeholder:text-muted-foreground"
-            />
-          </div>
-        </div>
-        <div className="max-h-60 overflow-y-auto p-1">
-          {!hasFilteredOptions ? (
-            <div className="px-3 py-4 text-center text-xs text-muted-foreground">
-              No models found
-            </div>
-          ) : (
-            <>
-              {filteredAutoOption ? (
-                <div className="pb-1">{renderOption(filteredAutoOption)}</div>
-              ) : null}
-              {filteredOptionGroups.map((group, idx) => (
-                <div
-                  key={group.label || "models"}
-                  className={idx > 0 ? "mt-2" : ""}
-                >
-                  {group.label ? (
-                    <div className="px-2.5 pb-1 text-[10px] font-medium uppercase text-muted-foreground">
-                      {group.label}
-                    </div>
-                  ) : null}
-                  {group.options.map((option) => renderOption(option))}
-                </div>
-              ))}
-            </>
-          )}
-        </div>
-      </PopoverContent>
-    </Popover>
-  );
-}
-
-function ThinkingValueSelect({
-  selectedThinkingValue,
-  thinkingValues,
-  disabled,
-  compact = false,
-  compactWidth,
-  onThinkingValueChange,
-}: {
-  selectedThinkingValue: string | null;
-  thinkingValues: string[];
-  disabled: boolean;
-  compact?: boolean;
-  compactWidth?: number;
-  onThinkingValueChange: (value: string | null) => void;
-}) {
-  const [open, setOpen] = useState(false);
-
-  if (thinkingValues.length === 0 || !selectedThinkingValue) {
-    return null;
-  }
-  const selectedThinkingLabel = displayThinkingValueLabel(
-    selectedThinkingValue,
-  );
-  const showCompactLabel = !compact || typeof compactWidth !== "number";
-
-  const renderOption = (value: string) => {
-    const active = value === selectedThinkingValue;
-    return (
-      <button
-        key={value}
-        type="button"
-        aria-current={active ? "true" : undefined}
-        onClick={() => {
-          onThinkingValueChange(value);
-          setOpen(false);
-        }}
-        className={`flex w-full items-center justify-between gap-2 rounded-md px-2.5 py-1.5 text-left text-xs transition-colors ${
-          active
-            ? "bg-accent text-foreground"
-            : "text-foreground hover:bg-accent"
-        }`}
-      >
-        <span className="truncate">{displayThinkingValueLabel(value)}</span>
-      </button>
-    );
-  };
-
-  return (
-    <Popover
-      open={open}
-      onOpenChange={(nextOpen) => {
-        setOpen(nextOpen);
-      }}
-    >
-      <PopoverTrigger
-        disabled={disabled}
-        render={
-          <Button
-            variant="ghost"
-            size="sm"
-            aria-label={
-              compact ? `Reasoning effort: ${selectedThinkingLabel}` : undefined
-            }
-            className={`gap-1.5 rounded-md text-xs font-medium ${
-              compact
-                ? showCompactLabel
-                  ? "w-full min-w-0 justify-between px-2.5"
-                  : "w-full min-w-0 justify-center px-2.5"
-                : "px-2"
-            }`}
-          >
-            {compact ? (
-              showCompactLabel ? (
-                <>
-                  <span className="flex min-w-0 items-center gap-1.5">
-                    <Lightbulb className="size-3.5 shrink-0 text-muted-foreground" />
-                    <span className="truncate">{selectedThinkingLabel}</span>
-                  </span>
-                  <ChevronDown className="size-3.5 shrink-0 text-muted-foreground" />
-                </>
-              ) : (
-                <span className="flex min-w-0 items-center gap-1.5">
-                  <Lightbulb className="size-3.5 shrink-0 text-muted-foreground" />
-                  <ChevronDown className="size-3.5 shrink-0 text-muted-foreground" />
-                </span>
-              )
-            ) : (
-              <>
-                <span className="whitespace-nowrap">
-                  {selectedThinkingLabel}
-                </span>
-                <ChevronDown className="size-3.5 shrink-0 text-muted-foreground" />
-              </>
-            )}
-          </Button>
-        }
-      />
-      <PopoverContent
-        align="start"
-        side="top"
-        sideOffset={8}
-        className="max-w-40 gap-0 rounded-lg p-1 shadow-xs ring-0"
-      >
-        <div className="px-2.5 pb-1 pt-1 text-[10px] font-medium uppercase text-muted-foreground">
-          Reasoning effort
-        </div>
-        {thinkingValues.map((value) => renderOption(value))}
-      </PopoverContent>
-    </Popover>
-  );
-}
-
-function Composer({
-  input,
-  quotedSkills,
-  slashCommands,
-  attachments,
-  isResponding,
-  pausePending,
-  pauseDisabled,
-  disabled,
-  disabledReason = "",
-  selectedModel,
-  resolvedModelLabel,
-  runtimeDefaultModelLabel,
-  modelOptions,
-  modelOptionGroups,
-  runtimeDefaultModelAvailable,
-  selectedThinkingValue,
-  thinkingValues,
-  showThinkingValueSelector,
-  modelSelectionUnavailableReason,
-  submitDisabled = false,
-  placeholder,
-  showModelSelector,
-  onModelChange,
-  onThinkingValueChange,
-  onOpenModelProviders,
-  textareaRef,
-  fileInputRef,
-  onChange,
-  onKeyDown,
-  onCompositionStart,
-  onCompositionEnd,
-  onAttachmentInputChange,
-  onPause,
-  onAddDroppedFiles,
-  onAddExplorerAttachments,
-  onSelectSlashCommand,
-  mentionableItems,
-  onRemoveQuotedSkill,
-  onRemoveAttachment,
-  onPreviewAttachment,
-}: ComposerProps) {
-  const [isDragActive, setIsDragActive] = useState(false);
-  const [composerActionsMenuOpen, setComposerActionsMenuOpen] = useState(false);
-  const [composerActionsView, setComposerActionsView] = useState<
-    "menu" | "skills"
-  >("menu");
-  const [skillPickerQuery, setSkillPickerQuery] = useState("");
-  const [caretIndex, setCaretIndex] = useState(0);
-  const [dismissedSlashCommandKey, setDismissedSlashCommandKey] = useState("");
-  const [highlightedSlashIndex, setHighlightedSlashIndex] = useState(0);
-  const [dismissedMentionKey, setDismissedMentionKey] = useState("");
-  const [highlightedMentionIndex, setHighlightedMentionIndex] = useState(0);
-  const composerFooterRef = useRef<HTMLDivElement | null>(null);
-  const composerActionsRef = useRef<HTMLDivElement | null>(null);
-  const composerFooterLayoutSyncFrameRef = useRef<number | null>(null);
-  const slashCommandMenuRef = useRef<HTMLDivElement | null>(null);
-  const [composerFooterLayout, setComposerFooterLayout] = useState({
-    width: 0,
-    actionsWidth: 0,
-  });
-  const noAvailableModels =
-    !runtimeDefaultModelAvailable &&
-    modelOptions.length === 0 &&
-    modelOptionGroups.length === 0;
-  const inputDisabled = disabled;
-  const activeSlashRange = useMemo(
-    () => findActiveSlashCommandRange(input, caretIndex),
-    [caretIndex, input],
-  );
-  const activeSlashCommandKey = activeSlashRange
-    ? `${activeSlashRange.start}:${activeSlashRange.end}:${activeSlashRange.query}`
-    : "";
-  const showSlashCommandMenu =
-    !inputDisabled &&
-    activeSlashRange !== null &&
-    activeSlashCommandKey !== dismissedSlashCommandKey;
-  const activeMentionRange = useMemo(
-    // Slash takes precedence — both pickers can't be live at once.
-    () =>
-      activeSlashRange ? null : findActiveMentionRange(input, caretIndex),
-    [activeSlashRange, caretIndex, input],
-  );
-  const activeMentionKey = activeMentionRange
-    ? `${activeMentionRange.start}:${activeMentionRange.end}:${activeMentionRange.query}`
-    : "";
-  const mentionItemsList = mentionableItems ?? [];
-  const filteredMentionItems = useMemo(() => {
-    if (inputDisabled || !activeMentionRange || mentionItemsList.length === 0) {
-      return [];
-    }
-    const query = activeMentionRange.query.trim().toLowerCase();
-    if (!query) {
-      return mentionItemsList;
-    }
-    return mentionItemsList.filter((item) => {
-      const haystack = [item.handle, ...(item.keywords ?? [])]
-        .join(" ")
-        .toLowerCase();
-      return haystack.includes(query);
-    });
-  }, [activeMentionRange, inputDisabled, mentionItemsList]);
-  const showMentionMenu =
-    !inputDisabled &&
-    activeMentionRange !== null &&
-    mentionItemsList.length > 0 &&
-    filteredMentionItems.length > 0 &&
-    activeMentionKey !== dismissedMentionKey;
-  const filteredSlashCommands = useMemo(() => {
-    if (inputDisabled || !activeSlashRange) {
-      return [];
-    }
-    const query = activeSlashRange.query.trim().toLowerCase();
-    if (!query) {
-      return slashCommands;
-    }
-    return slashCommands.filter(
-      (command) =>
-        command.command.toLowerCase().includes(query) ||
-        command.searchText.includes(query),
-    );
-  }, [activeSlashRange, inputDisabled, slashCommands]);
-  const filteredSkillCommands = useMemo(() => {
-    const query = skillPickerQuery.trim().toLowerCase();
-    if (!query) {
-      return slashCommands;
-    }
-    return slashCommands.filter(
-      (command) =>
-        command.command.toLowerCase().includes(query) ||
-        command.searchText.includes(query),
-    );
-  }, [skillPickerQuery, slashCommands]);
-  const quotedSkillIdSet = useMemo(
-    () => new Set(quotedSkills.map((skill) => skill.skillId)),
-    [quotedSkills],
-  );
-  const visibleModelOptions = modelOptionGroups.flatMap(
-    (group) => group.options,
-  );
-  const selectedModelOptionLabel =
-    visibleModelOptions.find((option) => option.value === selectedModel)
-      ?.selectedLabel ??
-    visibleModelOptions.find((option) => option.value === selectedModel)
-      ?.label ??
-    modelOptions.find((option) => option.value === selectedModel)
-      ?.selectedLabel ??
-    modelOptions.find((option) => option.value === selectedModel)?.label ??
-    resolvedModelLabel;
-  const cancelComposerFooterLayoutSync = () => {
-    if (composerFooterLayoutSyncFrameRef.current === null) {
-      return;
-    }
-    window.cancelAnimationFrame(composerFooterLayoutSyncFrameRef.current);
-    composerFooterLayoutSyncFrameRef.current = null;
-  };
-  const syncComposerFooterLayout = () => {
-    const footer = composerFooterRef.current;
-    if (!footer) {
-      return;
-    }
-    const footerStyle = window.getComputedStyle(footer);
-    const horizontalPadding =
-      Number.parseFloat(footerStyle.paddingLeft || "0") +
-      Number.parseFloat(footerStyle.paddingRight || "0");
-    const width = Math.max(
-      0,
-      Math.round(footer.clientWidth - horizontalPadding),
-    );
-    const actionsWidth = Math.round(
-      composerActionsRef.current?.getBoundingClientRect().width ?? 0,
-    );
-    setComposerFooterLayout((current) =>
-      current.width === width && current.actionsWidth === actionsWidth
-        ? current
-        : { width, actionsWidth },
-    );
-  };
-  // Coalesce ResizeObserver bursts so compact/full footer transitions do not
-  // synchronously re-enter render while the DOM is still settling.
-  const scheduleComposerFooterLayoutSync = () => {
-    if (composerFooterLayoutSyncFrameRef.current !== null) {
-      return;
-    }
-    composerFooterLayoutSyncFrameRef.current = window.requestAnimationFrame(
-      () => {
-        composerFooterLayoutSyncFrameRef.current = null;
-        syncComposerFooterLayout();
-      },
-    );
-  };
-  useLayoutEffect(() => {
-    const footer = composerFooterRef.current;
-    if (!footer) {
-      return;
-    }
-
-    syncComposerFooterLayout();
-    if (typeof ResizeObserver === "undefined") {
-      return;
-    }
-
-    const resizeObserver = new ResizeObserver(() => {
-      scheduleComposerFooterLayoutSync();
-    });
-    resizeObserver.observe(footer);
-    if (composerActionsRef.current) {
-      resizeObserver.observe(composerActionsRef.current);
-    }
-    return () => {
-      resizeObserver.disconnect();
-      cancelComposerFooterLayoutSync();
-    };
-  }, []);
-  useEffect(() => {
-    setHighlightedSlashIndex(0);
-  }, [activeSlashRange?.query, filteredSlashCommands.length]);
-  useEffect(() => {
-    setHighlightedMentionIndex(0);
-  }, [activeMentionRange?.query, filteredMentionItems.length]);
-  useEffect(() => {
-    if (!dismissedMentionKey) {
-      return;
-    }
-    if (!activeMentionKey) {
-      setDismissedMentionKey("");
-      return;
-    }
-    if (dismissedMentionKey !== activeMentionKey) {
-      setDismissedMentionKey("");
-    }
-  }, [activeMentionKey, dismissedMentionKey]);
-  useEffect(() => {
-    if (!dismissedSlashCommandKey) {
-      return;
-    }
-    if (!activeSlashCommandKey) {
-      setDismissedSlashCommandKey("");
-      return;
-    }
-    if (dismissedSlashCommandKey !== activeSlashCommandKey) {
-      setDismissedSlashCommandKey("");
-    }
-  }, [activeSlashCommandKey, dismissedSlashCommandKey]);
-  useEffect(() => {
-    if (inputDisabled) {
-      setComposerActionsMenuOpen(false);
-      setComposerActionsView("menu");
-      setSkillPickerQuery("");
-      setDismissedSlashCommandKey("");
-    }
-  }, [inputDisabled]);
-  useEffect(() => {
-    if (!showSlashCommandMenu) {
-      return;
-    }
-
-    const handlePointerDown = (event: PointerEvent) => {
-      const menu = slashCommandMenuRef.current;
-      const target = event.target;
-      if (!menu || !(target instanceof Node) || menu.contains(target)) {
-        return;
-      }
-      setDismissedSlashCommandKey(activeSlashCommandKey);
-    };
-
-    document.addEventListener("pointerdown", handlePointerDown);
-    return () => {
-      document.removeEventListener("pointerdown", handlePointerDown);
-    };
-  }, [activeSlashCommandKey, showSlashCommandMenu]);
-  const visibleFooterControlCount = 1 + (showThinkingValueSelector ? 1 : 0) + 1;
-  const fullPrimaryControlWidth = showModelSelector
-    ? noAvailableModels
-      ? COMPOSER_FULL_PROVIDER_SETUP_WIDTH_PX
-      : COMPOSER_FULL_MODEL_CONTROL_WIDTH_PX
-    : 0;
-  const fullFooterControlWidth =
-    fullPrimaryControlWidth +
-    (showThinkingValueSelector ? COMPOSER_FULL_THINKING_CONTROL_WIDTH_PX : 0) +
-    composerFooterLayout.actionsWidth +
-    Math.max(0, visibleFooterControlCount - 1) * COMPOSER_FOOTER_GAP_PX;
-  const compactFooterControlWidth = Math.max(
-    0,
-    composerFooterLayout.width -
-      composerFooterLayout.actionsWidth -
-      Math.max(0, visibleFooterControlCount - 1) * COMPOSER_FOOTER_GAP_PX,
-  );
-  const compactComposerControls =
-    showModelSelector &&
-    composerFooterLayout.width > 0 &&
-    composerFooterLayout.actionsWidth > 0 &&
-    composerFooterLayout.width < fullFooterControlWidth;
-  const compactModelControlWidth = compactComposerControls
-    ? Math.min(
-        COMPOSER_COMPACT_MODEL_CONTROL_MAX_WIDTH_PX,
-        Math.max(
-          0,
-          compactFooterControlWidth -
-            (showThinkingValueSelector
-              ? Math.min(
-                  COMPOSER_COMPACT_THINKING_CONTROL_MIN_WIDTH_PX,
-                  compactFooterControlWidth,
-                )
-              : 0),
-        ),
-      )
-    : 0;
-  const compactThinkingControlWidth = showThinkingValueSelector
-    ? Math.max(
-        Math.min(
-          COMPOSER_COMPACT_THINKING_CONTROL_MAX_WIDTH_PX,
-          compactFooterControlWidth - compactModelControlWidth,
-        ),
-        Math.min(
-          COMPOSER_COMPACT_THINKING_CONTROL_MIN_WIDTH_PX,
-          compactFooterControlWidth,
-        ),
-      )
-    : 0;
-
-  const syncCaretFromTextarea = (target: HTMLTextAreaElement | null) => {
-    if (!target) {
-      return;
-    }
-    setCaretIndex(target.selectionStart ?? target.value.length);
-  };
-
-  const handleTextareaChange = (event: ChangeEvent<HTMLTextAreaElement>) => {
-    onChange(event.target.value);
-    syncCaretFromTextarea(event.target);
-  };
-
-  const applySlashCommand = (command: ChatComposerSlashCommandOption) => {
-    onSelectSlashCommand(command);
-    if (!activeSlashRange) {
-      window.requestAnimationFrame(() => {
-        const textarea = textareaRef.current;
-        if (!textarea) {
-          return;
-        }
-        textarea.focus();
-        syncCaretFromTextarea(textarea);
-      });
-      return;
-    }
-    const nextInput = removeSlashCommandText(input, activeSlashRange);
-    onChange(nextInput.value);
-    setCaretIndex(nextInput.caretIndex);
-    window.requestAnimationFrame(() => {
-      const textarea = textareaRef.current;
-      if (!textarea) {
-        return;
-      }
-      textarea.focus();
-      textarea.setSelectionRange(nextInput.caretIndex, nextInput.caretIndex);
-    });
-  };
-
-  const applyMentionItem = (item: ChatComposerMentionItem) => {
-    if (!activeMentionRange) {
-      return;
-    }
-    const nextInput = replaceMentionText(
-      input,
-      activeMentionRange,
-      `@${item.handle}`,
-    );
-    onChange(nextInput.value);
-    setCaretIndex(nextInput.caretIndex);
-    setDismissedMentionKey("");
-    window.requestAnimationFrame(() => {
-      const textarea = textareaRef.current;
-      if (!textarea) {
-        return;
-      }
-      textarea.focus();
-      textarea.setSelectionRange(nextInput.caretIndex, nextInput.caretIndex);
-    });
-  };
-
-  const handleTextareaKeyDown = (event: KeyboardEvent<HTMLTextAreaElement>) => {
-    if (showMentionMenu) {
-      if (event.key === "ArrowDown" && filteredMentionItems.length > 0) {
-        event.preventDefault();
-        setHighlightedMentionIndex(
-          (current) => (current + 1) % filteredMentionItems.length,
-        );
-        return;
-      }
-      if (event.key === "ArrowUp" && filteredMentionItems.length > 0) {
-        event.preventDefault();
-        setHighlightedMentionIndex(
-          (current) =>
-            (current - 1 + filteredMentionItems.length) %
-            filteredMentionItems.length,
-        );
-        return;
-      }
-      if (
-        (event.key === "Enter" || event.key === "Tab") &&
-        filteredMentionItems.length > 0
-      ) {
-        event.preventDefault();
-        applyMentionItem(
-          filteredMentionItems[
-            Math.min(highlightedMentionIndex, filteredMentionItems.length - 1)
-          ]!,
-        );
-        return;
-      }
-      if (event.key === "Escape") {
-        event.preventDefault();
-        setDismissedMentionKey(activeMentionKey);
-        return;
-      }
-    }
-    if (showSlashCommandMenu) {
-      if (event.key === "ArrowDown" && filteredSlashCommands.length > 0) {
-        event.preventDefault();
-        setHighlightedSlashIndex(
-          (current) => (current + 1) % filteredSlashCommands.length,
-        );
-        return;
-      }
-      if (event.key === "ArrowUp" && filteredSlashCommands.length > 0) {
-        event.preventDefault();
-        setHighlightedSlashIndex(
-          (current) =>
-            (current - 1 + filteredSlashCommands.length) %
-            filteredSlashCommands.length,
-        );
-        return;
-      }
-      if (
-        (event.key === "Enter" || event.key === "Tab") &&
-        filteredSlashCommands.length > 0
-      ) {
-        event.preventDefault();
-        applySlashCommand(
-          filteredSlashCommands[
-            Math.min(highlightedSlashIndex, filteredSlashCommands.length - 1)
-          ]!,
-        );
-        return;
-      }
-      if (event.key === "Escape") {
-        event.preventDefault();
-        setCaretIndex(-1);
-        return;
-      }
-    }
-    onKeyDown(event);
-  };
-
-  const handleTextareaPaste = (event: ClipboardEvent<HTMLTextAreaElement>) => {
-    const pastedFiles = clipboardFilesFromDataTransfer(event.clipboardData);
-    if (pastedFiles.length === 0) {
-      const clipboardText =
-        event.clipboardData?.getData("text/plain")?.trim() ?? "";
-      const explorerFiles =
-        explorerAttachmentFilesFromClipboardText(clipboardText);
-      if (explorerFiles.length > 0) {
-        event.preventDefault();
-        onAddExplorerAttachments(explorerFiles);
-        return;
-      }
-
-      const clipboardTypes = Array.from(event.clipboardData?.types ?? []);
-      const hasClipboardImageType = clipboardTypes.some(
-        (type) => type === "Files" || type.startsWith("image/"),
-      );
-      if (
-        clipboardText ||
-        (clipboardTypes.includes("text/html") && !hasClipboardImageType)
-      ) {
-        return;
-      }
-
-      event.preventDefault();
-      void clipboardImageFileFromElectronClipboard()
-        .then((file) => {
-          if (file) {
-            onAddDroppedFiles([file]);
-          }
-        })
-        .catch(() => undefined);
-      return;
-    }
-
-    event.preventDefault();
-    onAddDroppedFiles(pastedFiles);
-  };
-
-  const openSkillPickerFromComposerMenu = () => {
-    setComposerActionsView("skills");
-    setSkillPickerQuery("");
-  };
-
-  const closeComposerActionsMenu = () => {
-    setComposerActionsMenuOpen(false);
-    setComposerActionsView("menu");
-    setSkillPickerQuery("");
-  };
-
-  const selectSkillFromPicker = (command: ChatComposerSlashCommandOption) => {
-    const alreadyQuoted = quotedSkills.some(
-      (skill) => skill.skillId === command.skillId,
-    );
-    if (alreadyQuoted) {
-      onRemoveQuotedSkill(command.skillId);
-      return;
-    }
-    onSelectSlashCommand(command);
-    window.requestAnimationFrame(() => {
-      const textarea = textareaRef.current;
-      if (!textarea) {
-        return;
-      }
-      textarea.focus();
-      syncCaretFromTextarea(textarea);
-    });
-  };
-
-  const allowAttachmentDrop = (dataTransfer: DataTransfer | null) => {
-    if (!dataTransfer || disabled) {
-      return false;
-    }
-
-    const types = Array.from(dataTransfer.types ?? []);
-    if (types.includes(EXPLORER_ATTACHMENT_DRAG_TYPE)) {
-      return true;
-    }
-
-    if ((dataTransfer.files?.length ?? 0) > 0) {
-      return true;
-    }
-
-    return Array.from(dataTransfer.items ?? []).some(
-      (item) => item.kind === "file",
-    );
-  };
-
-  const onDragOver = (event: DragEvent<HTMLDivElement>) => {
-    if (!allowAttachmentDrop(event.dataTransfer)) {
-      return;
-    }
-
-    event.preventDefault();
-    event.dataTransfer.dropEffect = "copy";
-    if (!isDragActive) {
-      setIsDragActive(true);
-    }
-  };
-
-  const onDragLeave = (event: DragEvent<HTMLDivElement>) => {
-    if (event.currentTarget.contains(event.relatedTarget as Node | null)) {
-      return;
-    }
-    setIsDragActive(false);
-  };
-
-  const onDrop = (event: DragEvent<HTMLDivElement>) => {
-    if (!allowAttachmentDrop(event.dataTransfer)) {
-      return;
-    }
-
-    event.preventDefault();
-    setIsDragActive(false);
-
-    const explorerFiles: ExplorerAttachmentDragPayload[] = [];
-    const rawExplorerPayload = event.dataTransfer.getData(
-      EXPLORER_ATTACHMENT_DRAG_TYPE,
-    );
-    const parsedExplorerPayload =
-      parseExplorerAttachmentDragPayload(rawExplorerPayload);
-    if (parsedExplorerPayload) {
-      explorerFiles.push(parsedExplorerPayload);
-    }
-
-    const droppedFiles = Array.from(event.dataTransfer.files ?? []);
-    if (explorerFiles.length > 0) {
-      onAddExplorerAttachments(explorerFiles);
-    }
-    if (droppedFiles.length > 0) {
-      onAddDroppedFiles(droppedFiles);
-    }
-  };
-
-  return (
-    <div className="relative">
-      {showSlashCommandMenu ? (
-        <div className="pointer-events-none absolute left-3 right-3 top-4 z-20 -translate-y-[calc(100%+2px)]">
-          <div
-            ref={slashCommandMenuRef}
-            className="pointer-events-auto overflow-hidden rounded-lg bg-popover shadow-md"
-          >
-            {filteredSlashCommands.length > 0 ? (
-              <div className="max-h-[280px] overflow-y-auto p-1">
-                {filteredSlashCommands.map((command, index) => (
-                  <button
-                    key={command.key}
-                    type="button"
-                    onClick={() => applySlashCommand(command)}
-                    className={`flex w-full items-center gap-2 rounded-md px-2 py-1 text-left text-xs transition-colors ${
-                      index === highlightedSlashIndex
-                        ? "bg-fg-8 text-foreground"
-                        : "hover:bg-fg-4"
-                    }`}
-                  >
-                    <Wand2 className="size-3.5 shrink-0 text-muted-foreground" />
-                    <span className="min-w-0 flex-1 truncate font-medium">
-                      {command.label}
-                    </span>
-                  </button>
-                ))}
-              </div>
-            ) : (
-              <div className="px-3 py-2.5 text-xs text-muted-foreground">
-                {slashCommands.length === 0
-                  ? "No skills in this workspace yet — add one to create slash commands."
-                  : "No skills match."}
-              </div>
-            )}
-          </div>
-        </div>
-      ) : null}
-      {showMentionMenu ? (
-        <div className="pointer-events-none absolute left-3 right-3 top-4 z-20 -translate-y-[calc(100%+2px)]">
-          <div className="pointer-events-auto overflow-hidden rounded-lg bg-popover shadow-md">
-            <div className="max-h-[280px] overflow-y-auto p-1">
-              {filteredMentionItems.map((item, index) => (
-                <button
-                  key={item.id}
-                  type="button"
-                  onClick={() => applyMentionItem(item)}
-                  className={`flex w-full items-center gap-2 rounded-md px-2 py-1 text-left text-xs transition-colors ${
-                    index === highlightedMentionIndex
-                      ? "bg-fg-8 text-foreground"
-                      : "hover:bg-fg-4"
-                  }`}
-                >
-                  {item.kindIcon ? (
-                    <span
-                      aria-hidden="true"
-                      className="shrink-0 text-muted-foreground"
-                    >
-                      {item.kindIcon}
-                    </span>
-                  ) : null}
-                  <span className="min-w-0 flex-1 truncate font-medium">
-                    {item.label}
-                  </span>
-                </button>
-              ))}
-            </div>
-          </div>
-        </div>
-      ) : null}
-
-      <div
-        onDragOver={onDragOver}
-        onDragLeave={onDragLeave}
-        onDrop={onDrop}
-        className="relative overflow-hidden rounded-2xl bg-background shadow-md"
-      >
-        {isDragActive ? (
-          <div className="pointer-events-none absolute inset-0 z-10 flex items-center justify-center rounded-2xl border border-dashed border-primary/50 bg-background/85 backdrop-blur-sm">
-            <div className="flex flex-col items-center gap-1.5 text-primary">
-              <Paperclip className="size-5" />
-              <span className="text-xs font-medium">
-                Drop files to attach
-              </span>
-            </div>
-          </div>
-        ) : null}
-        <input
-          ref={fileInputRef}
-          type="file"
-          multiple
-          className="hidden"
-          onChange={onAttachmentInputChange}
-        />
-        {attachments.length > 0 ? (
-          <div className="border-b border-border px-4 py-3">
-            <AttachmentList
-              attachments={attachments}
-              onPreview={onPreviewAttachment}
-              onRemove={onRemoveAttachment}
-            />
-          </div>
-        ) : null}
-        {quotedSkills.length > 0 ? (
-          <div className="border-b border-border px-4 py-2.5">
-            <div className="flex flex-wrap gap-1">
-              {quotedSkills.map((skill) => (
-                <EntityChip
-                  key={skill.skillId}
-                  size="md"
-                  icon={<Sparkles className="text-muted-foreground" />}
-                  label={skill.title}
-                  trailing={
-                    <button
-                      type="button"
-                      onClick={() => onRemoveQuotedSkill(skill.skillId)}
-                      className="grid size-4 place-items-center rounded-sm text-muted-foreground transition hover:bg-fg-8 hover:text-foreground"
-                      aria-label={`Remove quoted skill ${skill.title}`}
-                    >
-                      <X className="size-3" />
-                    </button>
-                  }
-                />
-              ))}
-            </div>
-          </div>
-        ) : null}
-        <div className="px-5 pb-2 pt-3">
-          <textarea
-            ref={textareaRef}
-            value={input}
-            onChange={handleTextareaChange}
-            onKeyDown={handleTextareaKeyDown}
-            onPaste={handleTextareaPaste}
-            onSelect={(event) => syncCaretFromTextarea(event.currentTarget)}
-            onClick={(event) => syncCaretFromTextarea(event.currentTarget)}
-            onCompositionStart={onCompositionStart}
-            onCompositionEnd={onCompositionEnd}
-            rows={1}
-            disabled={inputDisabled}
-            placeholder={
-              inputDisabled
-                ? disabledReason || "Chat unavailable right now"
-                : placeholder
-            }
-            className="composer-input block max-h-[220px] min-h-[40px] w-full resize-none overflow-y-auto bg-transparent text-sm leading-7 text-foreground outline-none placeholder:text-muted-foreground/50 disabled:cursor-not-allowed disabled:opacity-55"
-          />
-        </div>
-
-        <div
-          ref={composerFooterRef}
-          className={`px-3 pb-3 text-muted-foreground ${
-            compactComposerControls
-              ? "flex items-center gap-1.5 overflow-hidden"
-              : "flex flex-wrap items-center gap-1.5"
-          }`}
-        >
-          {showModelSelector ? (
-            <div
-              className={
-                compactComposerControls
-                  ? "min-w-0 shrink-0"
-                  : noAvailableModels
-                    ? "min-w-0 flex flex-1 basis-full flex-wrap items-center gap-2"
-                    : "min-w-0 shrink-0"
-              }
-            >
-              {noAvailableModels ? (
-                <>
-                  <Button
-                    type="button"
-                    variant="outline"
-                    size="lg"
-                    onClick={onOpenModelProviders}
-                    className={`shrink-0 justify-between rounded-lg bg-card text-xs font-semibold hover:border-primary hover:bg-card ${
-                      compactComposerControls ? "px-2.5" : ""
-                    }`}
-                    aria-label="Configure model providers"
-                  >
-                    <span className="flex min-w-0 items-center gap-2">
-                      <Waypoints className="size-3.5 shrink-0 text-muted-foreground" />
-                      <span className="truncate">
-                        {compactComposerControls
-                          ? "Providers"
-                          : "Set up providers"}
-                      </span>
-                    </span>
-                    <ArrowRight className="size-3.5 shrink-0 text-muted-foreground" />
-                  </Button>
-                  <div
-                    className={`min-w-0 text-[10px] leading-5 text-muted-foreground ${
-                      compactComposerControls ? "hidden" : ""
-                    }`}
-                  >
-                    Open provider settings to connect a model.
-                  </div>
-                </>
-              ) : (
-                <ModelCombobox
-                  selectedModel={selectedModel}
-                  selectedModelLabel={selectedModelOptionLabel}
-                  runtimeDefaultModelLabel={runtimeDefaultModelLabel}
-                  runtimeDefaultModelAvailable={runtimeDefaultModelAvailable}
-                  modelOptions={modelOptions}
-                  modelOptionGroups={modelOptionGroups}
-                  disabled={disabled}
-                  compact={compactComposerControls}
-                  onModelChange={onModelChange}
-                />
-              )}
-            </div>
-          ) : (
-            <div className="min-w-0 flex-1 text-xs leading-6 text-muted-foreground">
-              Responses here stay in the workspace onboarding thread.
-            </div>
-          )}
-
-          {showThinkingValueSelector ? (
-            <div className="shrink-0">
-              <ThinkingValueSelect
-                selectedThinkingValue={selectedThinkingValue}
-                thinkingValues={thinkingValues}
-                disabled={disabled}
-                compact={compactComposerControls}
-                compactWidth={
-                  compactComposerControls
-                    ? compactThinkingControlWidth
-                    : undefined
-                }
-                onThinkingValueChange={onThinkingValueChange}
-              />
-            </div>
-          ) : null}
-
-          <div
-            ref={composerActionsRef}
-            className="ml-auto flex shrink-0 items-center gap-1.5"
-          >
-            <Popover
-              open={composerActionsMenuOpen}
-              onOpenChange={(nextOpen) => {
-                setComposerActionsMenuOpen(nextOpen);
-                if (!nextOpen) {
-                  setComposerActionsView("menu");
-                  setSkillPickerQuery("");
-                }
-              }}
-            >
-              <PopoverTrigger
-                disabled={inputDisabled}
-                render={
-                  <Button
-                    variant="outline"
-                    size="icon-sm"
-                    aria-label="Open composer actions"
-                    className="rounded-lg"
-                  />
-                }
-              >
-                <Plus className="size-3.5" />
-              </PopoverTrigger>
-              <PopoverContent
-                align="end"
-                side="top"
-                sideOffset={8}
-                className={`gap-0 rounded-xl border border-border bg-popover p-0 shadow-xs ring-0 ${
-                  composerActionsView === "skills" ? "w-[320px]" : "w-[224px]"
-                }`}
-              >
-                {composerActionsView === "skills" ? (
-                  <div className="flex flex-col">
-                    <div className="flex items-center gap-1.5 border-b border-border px-2 py-1.5">
-                      <button
-                        type="button"
-                        onClick={() => {
-                          setComposerActionsView("menu");
-                          setSkillPickerQuery("");
-                        }}
-                        aria-label="Back to actions"
-                        className="flex size-6 shrink-0 items-center justify-center rounded-md text-muted-foreground transition-colors hover:bg-accent hover:text-foreground"
-                      >
-                        <ArrowLeft className="size-3.5" />
-                      </button>
-                      <Search className="size-3.5 shrink-0 text-muted-foreground" />
-                      <input
-                        value={skillPickerQuery}
-                        onChange={(event) =>
-                          setSkillPickerQuery(event.target.value)
-                        }
-                        placeholder="Search skills"
-                        className="embedded-input h-7 w-full min-w-0 bg-transparent text-xs text-foreground outline-none placeholder:text-muted-foreground"
-                        autoFocus
-                      />
-                      {quotedSkillIdSet.size > 0 ? (
-                        <span className="shrink-0 rounded-full bg-accent px-1.5 py-0.5 text-[10px] font-medium tabular-nums text-muted-foreground">
-                          {quotedSkillIdSet.size}
-                        </span>
-                      ) : null}
-                    </div>
-                    {filteredSkillCommands.length > 0 ? (
-                      <div className="max-h-[288px] overflow-y-auto p-1">
-                        {filteredSkillCommands.map((command) => {
-                          const isSelected = quotedSkillIdSet.has(
-                            command.skillId,
-                          );
-                          return (
-                            <button
-                              key={command.key}
-                              type="button"
-                              onClick={() => selectSkillFromPicker(command)}
-                              aria-pressed={isSelected}
-                              className={`group flex w-full items-start gap-2.5 rounded-md px-2 py-1.5 text-left text-xs transition-colors ${
-                                isSelected
-                                  ? "bg-accent text-foreground"
-                                  : "text-foreground hover:bg-accent"
-                              }`}
-                            >
-                              <span className="mt-0.5 flex size-4 shrink-0 items-center justify-center">
-                                {isSelected ? (
-                                  <Check className="size-3.5 text-primary" />
-                                ) : (
-                                  <Sparkles className="size-3.5 text-muted-foreground" />
-                                )}
-                              </span>
-                              <span className="min-w-0 flex-1">
-                                <span className="block truncate font-medium text-foreground">
-                                  {command.label}
-                                </span>
-                                {command.description ? (
-                                  <span className="mt-0.5 block truncate text-[11px] leading-4 text-muted-foreground">
-                                    {command.description}
-                                  </span>
-                                ) : null}
-                              </span>
-                            </button>
-                          );
-                        })}
-                      </div>
-                    ) : (
-                      <div className="flex flex-col items-center gap-1 px-4 py-6 text-center">
-                        <Sparkles className="size-4 text-muted-foreground" />
-                        <span className="text-xs font-medium text-foreground">
-                          No matching skills
-                        </span>
-                        <span className="text-[11px] leading-4 text-muted-foreground">
-                          Try a different search term.
-                        </span>
-                      </div>
-                    )}
-                  </div>
-                ) : (
-                  <div className="flex flex-col gap-0.5 p-1">
-                    <button
-                      type="button"
-                      onClick={() => {
-                        closeComposerActionsMenu();
-                        fileInputRef.current?.click();
-                      }}
-                      className="flex w-full items-center gap-2.5 rounded-md px-2 py-1.5 text-left text-xs text-foreground transition-colors hover:bg-accent"
-                    >
-                      <Paperclip className="size-3.5 shrink-0 text-muted-foreground" />
-                      <span className="min-w-0 flex-1 truncate">
-                        Attach a file
-                      </span>
-                    </button>
-                    <button
-                      type="button"
-                      onClick={openSkillPickerFromComposerMenu}
-                      className="flex w-full items-center gap-2.5 rounded-md px-2 py-1.5 text-left text-xs text-foreground transition-colors hover:bg-accent"
-                    >
-                      <Zap className="size-3.5 shrink-0 text-muted-foreground" />
-                      <span className="min-w-0 flex-1 truncate">
-                        Use Skills
-                      </span>
-                      {quotedSkillIdSet.size > 0 ? (
-                        <span className="shrink-0 rounded-full bg-accent px-1.5 py-0.5 text-[10px] font-medium tabular-nums text-muted-foreground">
-                          {quotedSkillIdSet.size}
-                        </span>
-                      ) : null}
-                      <ChevronRight className="size-3.5 shrink-0 text-muted-foreground" />
-                    </button>
-                  </div>
-                )}
-              </PopoverContent>
-            </Popover>
-            {isResponding ? (
-              <Button
-                type="button"
-                size="icon-sm"
-                aria-label="Pause"
-                disabled={pausePending || pauseDisabled || disabled}
-                onClick={onPause}
-                className="rounded-lg"
-              >
-                {pausePending ? (
-                  <Loader2 className="size-3.5 animate-spin" />
-                ) : (
-                  <Square className="size-3 fill-current" />
-                )}
-              </Button>
-            ) : (
-              <Button
-                size="icon-sm"
-                aria-label="Send message"
-                disabled={
-                  (!input.trim() &&
-                    attachments.length === 0 &&
-                    quotedSkills.length === 0) ||
-                  disabled ||
-                  submitDisabled
-                }
-                render={<button type="submit" />}
-                className="rounded-lg"
-              >
-                <ArrowUp className="size-3.5" />
-              </Button>
-            )}
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
